--- legacy/index.js
@@ -1,4 +1,0 @@
-
-var ls = require('./lib/lookup_service.js');
-
-module.exports = ls;
