--- legacy/lib/data/region_name_data.js
@@ -1,4503 +1,0 @@
-
-/* jshint indent: 2 */
-// Do not modify this file, it was generated by
-// ./tools/gen_time_zone.rb
-module.exports = {
-  "AD": {
-    "2": "Canillo",
-    "3": "Encamp",
-    "4": "La Massana",
-    "5": "Ordino",
-    "6": "Sant Julia de Loria",
-    "7": "Andorra la Vella",
-    "8": "Escaldes-Engordany"
-  },
-  "AE": {
-    "1": "Abu Dhabi",
-    "2": "Ajman",
-    "3": "Dubai",
-    "4": "Fujairah",
-    "5": "Ras Al Khaimah",
-    "6": "Sharjah",
-    "7": "Umm Al Quwain"
-  },
-  "AF": {
-    "1": "Badakhshan",
-    "2": "Badghis",
-    "3": "Baghlan",
-    "5": "Bamian",
-    "6": "Farah",
-    "7": "Faryab",
-    "8": "Ghazni",
-    "9": "Ghowr",
-    "10": "Helmand",
-    "11": "Herat",
-    "13": "Kabol",
-    "14": "Kapisa",
-    "17": "Lowgar",
-    "18": "Nangarhar",
-    "19": "Nimruz",
-    "23": "Kandahar",
-    "24": "Kondoz",
-    "26": "Takhar",
-    "27": "Vardak",
-    "28": "Zabol",
-    "29": "Paktika",
-    "30": "Balkh",
-    "31": "Jowzjan",
-    "32": "Samangan",
-    "33": "Sar-e Pol",
-    "34": "Konar",
-    "35": "Laghman",
-    "36": "Paktia",
-    "37": "Khowst",
-    "38": "Nurestan",
-    "39": "Oruzgan",
-    "40": "Parvan",
-    "41": "Daykondi",
-    "42": "Panjshir"
-  },
-  "AG": {
-    "1": "Barbuda",
-    "3": "Saint George",
-    "4": "Saint John",
-    "5": "Saint Mary",
-    "6": "Saint Paul",
-    "7": "Saint Peter",
-    "8": "Saint Philip",
-    "9": "Redonda"
-  },
-  "AL": {
-    "40": "Berat",
-    "41": "Diber",
-    "42": "Durres",
-    "43": "Elbasan",
-    "44": "Fier",
-    "45": "Gjirokaster",
-    "46": "Korce",
-    "47": "Kukes",
-    "48": "Lezhe",
-    "49": "Shkoder",
-    "50": "Tirane",
-    "51": "Vlore"
-  },
-  "AM": {
-    "1": "Aragatsotn",
-    "2": "Ararat",
-    "3": "Armavir",
-    "4": "Geghark'unik'",
-    "5": "Kotayk'",
-    "6": "Lorri",
-    "7": "Shirak",
-    "8": "Syunik'",
-    "9": "Tavush",
-    "10": "Vayots' Dzor",
-    "11": "Yerevan"
-  },
-  "AO": {
-    "1": "Benguela",
-    "2": "Bie",
-    "3": "Cabinda",
-    "4": "Cuando Cubango",
-    "5": "Cuanza Norte",
-    "6": "Cuanza Sul",
-    "7": "Cunene",
-    "8": "Huambo",
-    "9": "Huila",
-    "12": "Malanje",
-    "13": "Namibe",
-    "14": "Moxico",
-    "15": "Uige",
-    "16": "Zaire",
-    "17": "Lunda Norte",
-    "18": "Lunda Sul",
-    "19": "Bengo",
-    "20": "Luanda"
-  },
-  "AR": {
-    "1": "Buenos Aires",
-    "2": "Catamarca",
-    "3": "Chaco",
-    "4": "Chubut",
-    "5": "Cordoba",
-    "6": "Corrientes",
-    "7": "Distrito Federal",
-    "8": "Entre Rios",
-    "9": "Formosa",
-    "10": "Jujuy",
-    "11": "La Pampa",
-    "12": "La Rioja",
-    "13": "Mendoza",
-    "14": "Misiones",
-    "15": "Neuquen",
-    "16": "Rio Negro",
-    "17": "Salta",
-    "18": "San Juan",
-    "19": "San Luis",
-    "20": "Santa Cruz",
-    "21": "Santa Fe",
-    "22": "Santiago del Estero",
-    "23": "Tierra del Fuego",
-    "24": "Tucuman"
-  },
-  "AT": {
-    "1": "Burgenland",
-    "2": "Karnten",
-    "3": "Niederosterreich",
-    "4": "Oberosterreich",
-    "5": "Salzburg",
-    "6": "Steiermark",
-    "7": "Tirol",
-    "8": "Vorarlberg",
-    "9": "Wien"
-  },
-  "AU": {
-    "1": "Australian Capital Territory",
-    "2": "New South Wales",
-    "3": "Northern Territory",
-    "4": "Queensland",
-    "5": "South Australia",
-    "6": "Tasmania",
-    "7": "Victoria",
-    "8": "Western Australia"
-  },
-  "AZ": {
-    "1": "Abseron",
-    "2": "Agcabadi",
-    "3": "Agdam",
-    "4": "Agdas",
-    "5": "Agstafa",
-    "6": "Agsu",
-    "7": "Ali Bayramli",
-    "8": "Astara",
-    "9": "Baki",
-    "10": "Balakan",
-    "11": "Barda",
-    "12": "Beylaqan",
-    "13": "Bilasuvar",
-    "14": "Cabrayil",
-    "15": "Calilabad",
-    "16": "Daskasan",
-    "17": "Davaci",
-    "18": "Fuzuli",
-    "19": "Gadabay",
-    "20": "Ganca",
-    "21": "Goranboy",
-    "22": "Goycay",
-    "23": "Haciqabul",
-    "24": "Imisli",
-    "25": "Ismayilli",
-    "26": "Kalbacar",
-    "27": "Kurdamir",
-    "28": "Lacin",
-    "29": "Lankaran",
-    "30": "Lankaran",
-    "31": "Lerik",
-    "32": "Masalli",
-    "33": "Mingacevir",
-    "34": "Naftalan",
-    "35": "Naxcivan",
-    "36": "Neftcala",
-    "37": "Oguz",
-    "38": "Qabala",
-    "39": "Qax",
-    "40": "Qazax",
-    "41": "Qobustan",
-    "42": "Quba",
-    "43": "Qubadli",
-    "44": "Qusar",
-    "45": "Saatli",
-    "46": "Sabirabad",
-    "47": "Saki",
-    "48": "Saki",
-    "49": "Salyan",
-    "50": "Samaxi",
-    "51": "Samkir",
-    "52": "Samux",
-    "53": "Siyazan",
-    "54": "Sumqayit",
-    "55": "Susa",
-    "56": "Susa",
-    "57": "Tartar",
-    "58": "Tovuz",
-    "59": "Ucar",
-    "60": "Xacmaz",
-    "61": "Xankandi",
-    "62": "Xanlar",
-    "63": "Xizi",
-    "64": "Xocali",
-    "65": "Xocavand",
-    "66": "Yardimli",
-    "67": "Yevlax",
-    "68": "Yevlax",
-    "69": "Zangilan",
-    "70": "Zaqatala",
-    "71": "Zardab"
-  },
-  "BA": {
-    "1": "Federation of Bosnia and Herzegovina",
-    "3": "Brcko District",
-    "2": "Republika Srpska"
-  },
-  "BB": {
-    "1": "Christ Church",
-    "2": "Saint Andrew",
-    "3": "Saint George",
-    "4": "Saint James",
-    "5": "Saint John",
-    "6": "Saint Joseph",
-    "7": "Saint Lucy",
-    "8": "Saint Michael",
-    "9": "Saint Peter",
-    "10": "Saint Philip",
-    "11": "Saint Thomas"
-  },
-  "BD": {
-    "81": "Dhaka",
-    "82": "Khulna",
-    "83": "Rajshahi",
-    "84": "Chittagong",
-    "85": "Barisal",
-    "86": "Sylhet",
-    "87": "Rangpur"
-  },
-  "BE": {
-    "1": "Antwerpen",
-    "3": "Hainaut",
-    "4": "Liege",
-    "5": "Limburg",
-    "6": "Luxembourg",
-    "7": "Namur",
-    "8": "Oost-Vlaanderen",
-    "9": "West-Vlaanderen",
-    "10": "Brabant Wallon",
-    "11": "Brussels Hoofdstedelijk Gewest",
-    "12": "Vlaams-Brabant",
-    "13": "Flanders",
-    "14": "Wallonia"
-  },
-  "BF": {
-    "15": "Bam",
-    "19": "Boulkiemde",
-    "20": "Ganzourgou",
-    "21": "Gnagna",
-    "28": "Kouritenga",
-    "33": "Oudalan",
-    "34": "Passore",
-    "36": "Sanguie",
-    "40": "Soum",
-    "42": "Tapoa",
-    "44": "Zoundweogo",
-    "45": "Bale",
-    "46": "Banwa",
-    "47": "Bazega",
-    "48": "Bougouriba",
-    "49": "Boulgou",
-    "50": "Gourma",
-    "51": "Houet",
-    "52": "Ioba",
-    "53": "Kadiogo",
-    "54": "Kenedougou",
-    "55": "Komoe",
-    "56": "Komondjari",
-    "57": "Kompienga",
-    "58": "Kossi",
-    "59": "Koulpelogo",
-    "60": "Kourweogo",
-    "61": "Leraba",
-    "62": "Loroum",
-    "63": "Mouhoun",
-    "64": "Namentenga",
-    "65": "Naouri",
-    "66": "Nayala",
-    "67": "Noumbiel",
-    "68": "Oubritenga",
-    "69": "Poni",
-    "70": "Sanmatenga",
-    "71": "Seno",
-    "72": "Sissili",
-    "73": "Sourou",
-    "74": "Tuy",
-    "75": "Yagha",
-    "76": "Yatenga",
-    "77": "Ziro",
-    "78": "Zondoma"
-  },
-  "BG": {
-    "33": "Mikhaylovgrad",
-    "38": "Blagoevgrad",
-    "39": "Burgas",
-    "40": "Dobrich",
-    "41": "Gabrovo",
-    "42": "Grad Sofiya",
-    "43": "Khaskovo",
-    "44": "Kurdzhali",
-    "45": "Kyustendil",
-    "46": "Lovech",
-    "47": "Montana",
-    "48": "Pazardzhik",
-    "49": "Pernik",
-    "50": "Pleven",
-    "51": "Plovdiv",
-    "52": "Razgrad",
-    "53": "Ruse",
-    "54": "Shumen",
-    "55": "Silistra",
-    "56": "Sliven",
-    "57": "Smolyan",
-    "58": "Sofiya",
-    "59": "Stara Zagora",
-    "60": "Turgovishte",
-    "61": "Varna",
-    "62": "Veliko Turnovo",
-    "63": "Vidin",
-    "64": "Vratsa",
-    "65": "Yambol"
-  },
-  "BH": {
-    "1": "Al Hadd",
-    "2": "Al Manamah",
-    "5": "Jidd Hafs",
-    "6": "Sitrah",
-    "8": "Al Mintaqah al Gharbiyah",
-    "9": "Mintaqat Juzur Hawar",
-    "10": "Al Mintaqah ash Shamaliyah",
-    "11": "Al Mintaqah al Wusta",
-    "12": "Madinat",
-    "13": "Ar Rifa",
-    "14": "Madinat Hamad",
-    "15": "Al Muharraq",
-    "16": "Al Asimah",
-    "17": "Al Janubiyah",
-    "18": "Ash Shamaliyah",
-    "19": "Al Wusta"
-  },
-  "BI": {
-    "2": "Bujumbura",
-    "9": "Bubanza",
-    "10": "Bururi",
-    "11": "Cankuzo",
-    "12": "Cibitoke",
-    "13": "Gitega",
-    "14": "Karuzi",
-    "15": "Kayanza",
-    "16": "Kirundo",
-    "17": "Makamba",
-    "18": "Muyinga",
-    "19": "Ngozi",
-    "20": "Rutana",
-    "21": "Ruyigi",
-    "22": "Muramvya",
-    "23": "Mwaro"
-  },
-  "BJ": {
-    "7": "Alibori",
-    "8": "Atakora",
-    "9": "Atlanyique",
-    "10": "Borgou",
-    "11": "Collines",
-    "12": "Kouffo",
-    "13": "Donga",
-    "14": "Littoral",
-    "15": "Mono",
-    "16": "Oueme",
-    "17": "Plateau",
-    "18": "Zou"
-  },
-  "BM": {
-    "1": "Devonshire",
-    "2": "Hamilton",
-    "3": "Hamilton",
-    "4": "Paget",
-    "5": "Pembroke",
-    "6": "Saint George",
-    "7": "Saint George's",
-    "8": "Sandys",
-    "9": "Smiths",
-    "10": "Southampton",
-    "11": "Warwick"
-  },
-  "BN": {
-    "7": "Alibori",
-    "8": "Belait",
-    "9": "Brunei and Muara",
-    "10": "Temburong",
-    "11": "Collines",
-    "12": "Kouffo",
-    "13": "Donga",
-    "14": "Littoral",
-    "15": "Tutong",
-    "16": "Oueme",
-    "17": "Plateau",
-    "18": "Zou"
-  },
-  "BO": {
-    "1": "Chuquisaca",
-    "2": "Cochabamba",
-    "3": "El Beni",
-    "4": "La Paz",
-    "5": "Oruro",
-    "6": "Pando",
-    "7": "Potosi",
-    "8": "Santa Cruz",
-    "9": "Tarija"
-  },
-  "BR": {
-    "1": "Acre",
-    "2": "Alagoas",
-    "3": "Amapa",
-    "4": "Amazonas",
-    "5": "Bahia",
-    "6": "Ceara",
-    "7": "Distrito Federal",
-    "8": "Espirito Santo",
-    "11": "Mato Grosso do Sul",
-    "13": "Maranhao",
-    "14": "Mato Grosso",
-    "15": "Minas Gerais",
-    "16": "Para",
-    "17": "Paraiba",
-    "18": "Parana",
-    "20": "Piaui",
-    "21": "Rio de Janeiro",
-    "22": "Rio Grande do Norte",
-    "23": "Rio Grande do Sul",
-    "24": "Rondonia",
-    "25": "Roraima",
-    "26": "Santa Catarina",
-    "27": "Sao Paulo",
-    "28": "Sergipe",
-    "29": "Goias",
-    "30": "Pernambuco",
-    "31": "Tocantins"
-  },
-  "BS": {
-    "5": "Bimini",
-    "6": "Cat Island",
-    "10": "Exuma",
-    "13": "Inagua",
-    "15": "Long Island",
-    "16": "Mayaguana",
-    "18": "Ragged Island",
-    "22": "Harbour Island",
-    "23": "New Providence",
-    "24": "Acklins and Crooked Islands",
-    "25": "Freeport",
-    "26": "Fresh Creek",
-    "27": "Governor's Harbour",
-    "28": "Green Turtle Cay",
-    "29": "High Rock",
-    "30": "Kemps Bay",
-    "31": "Marsh Harbour",
-    "32": "Nichollstown and Berry Islands",
-    "33": "Rock Sound",
-    "34": "Sandy Point",
-    "35": "San Salvador and Rum Cay"
-  },
-  "BT": {
-    "5": "Bumthang",
-    "6": "Chhukha",
-    "7": "Chirang",
-    "8": "Daga",
-    "9": "Geylegphug",
-    "10": "Ha",
-    "11": "Lhuntshi",
-    "12": "Mongar",
-    "13": "Paro",
-    "14": "Pemagatsel",
-    "15": "Punakha",
-    "16": "Samchi",
-    "17": "Samdrup",
-    "18": "Shemgang",
-    "19": "Tashigang",
-    "20": "Thimphu",
-    "21": "Tongsa",
-    "22": "Wangdi Phodrang"
-  },
-  "BW": {
-    "1": "Central",
-    "3": "Ghanzi",
-    "4": "Kgalagadi",
-    "5": "Kgatleng",
-    "6": "Kweneng",
-    "8": "North-East",
-    "9": "South-East",
-    "10": "Southern",
-    "11": "North-West"
-  },
-  "BY": {
-    "1": "Brestskaya Voblasts'",
-    "2": "Homyel'skaya Voblasts'",
-    "3": "Hrodzyenskaya Voblasts'",
-    "4": "Minsk",
-    "5": "Minskaya Voblasts'",
-    "6": "Mahilyowskaya Voblasts'",
-    "7": "Vitsyebskaya Voblasts'"
-  },
-  "BZ": {
-    "1": "Belize",
-    "2": "Cayo",
-    "3": "Corozal",
-    "4": "Orange Walk",
-    "5": "Stann Creek",
-    "6": "Toledo"
-  },
-  "CA": {
-    "849": "Alberta",
-    "893": "British Columbia",
-    "1365": "Manitoba",
-    "1408": "New Brunswick",
-    "1418": "Newfoundland",
-    "1425": "Nova Scotia",
-    "1426": "Northwest Territories",
-    "1427": "Nunavut",
-    "1463": "Ontario",
-    "1497": "Prince Edward Island",
-    "1538": "Quebec",
-    "1632": "Saskatchewan",
-    "1899": "Yukon Territory"
-  },
-  "CD": {
-    "1": "Bandundu",
-    "2": "Equateur",
-    "4": "Kasai-Oriental",
-    "5": "Katanga",
-    "6": "Kinshasa",
-    "8": "Bas-Congo",
-    "9": "Orientale",
-    "10": "Maniema",
-    "11": "Nord-Kivu",
-    "12": "Sud-Kivu"
-  },
-  "CF": {
-    "1": "Bamingui-Bangoran",
-    "2": "Basse-Kotto",
-    "3": "Haute-Kotto",
-    "4": "Mambere-Kadei",
-    "5": "Haut-Mbomou",
-    "6": "Kemo",
-    "7": "Lobaye",
-    "8": "Mbomou",
-    "9": "Nana-Mambere",
-    "11": "Ouaka",
-    "12": "Ouham",
-    "13": "Ouham-Pende",
-    "14": "Cuvette-Ouest",
-    "15": "Nana-Grebizi",
-    "16": "Sangha-Mbaere",
-    "17": "Ombella-Mpoko",
-    "18": "Bangui"
-  },
-  "CG": {
-    "1": "Bouenza",
-    "4": "Kouilou",
-    "5": "Lekoumou",
-    "6": "Likouala",
-    "7": "Niari",
-    "8": "Plateaux",
-    "10": "Sangha",
-    "11": "Pool",
-    "12": "Brazzaville",
-    "13": "Cuvette",
-    "14": "Cuvette-Ouest"
-  },
-  "CH": {
-    "1": "Aargau",
-    "2": "Ausser-Rhoden",
-    "3": "Basel-Landschaft",
-    "4": "Basel-Stadt",
-    "5": "Bern",
-    "6": "Fribourg",
-    "7": "Geneve",
-    "8": "Glarus",
-    "9": "Graubunden",
-    "10": "Inner-Rhoden",
-    "11": "Luzern",
-    "12": "Neuchatel",
-    "13": "Nidwalden",
-    "14": "Obwalden",
-    "15": "Sankt Gallen",
-    "16": "Schaffhausen",
-    "17": "Schwyz",
-    "18": "Solothurn",
-    "19": "Thurgau",
-    "20": "Ticino",
-    "21": "Uri",
-    "22": "Valais",
-    "23": "Vaud",
-    "24": "Zug",
-    "25": "Zurich",
-    "26": "Jura"
-  },
-  "CI": {
-    "74": "Agneby",
-    "75": "Bafing",
-    "76": "Bas-Sassandra",
-    "77": "Denguele",
-    "78": "Dix-Huit Montagnes",
-    "79": "Fromager",
-    "80": "Haut-Sassandra",
-    "81": "Lacs",
-    "82": "Lagunes",
-    "83": "Marahoue",
-    "84": "Moyen-Cavally",
-    "85": "Moyen-Comoe",
-    "86": "N'zi-Comoe",
-    "87": "Savanes",
-    "88": "Sud-Bandama",
-    "89": "Sud-Comoe",
-    "90": "Vallee du Bandama",
-    "91": "Worodougou",
-    "92": "Zanzan"
-  },
-  "CL": {
-    "1": "Valparaiso",
-    "2": "Aisen del General Carlos Ibanez del Campo",
-    "3": "Antofagasta",
-    "4": "Araucania",
-    "5": "Atacama",
-    "6": "Bio-Bio",
-    "7": "Coquimbo",
-    "8": "Libertador General Bernardo O'Higgins",
-    "9": "Los Lagos",
-    "10": "Magallanes y de la Antartica Chilena",
-    "11": "Maule",
-    "12": "Region Metropolitana",
-    "13": "Tarapaca",
-    "14": "Los Lagos",
-    "15": "Tarapaca",
-    "16": "Arica y Parinacota",
-    "17": "Los Rios"
-  },
-  "CM": {
-    "4": "Est",
-    "5": "Littoral",
-    "7": "Nord-Ouest",
-    "8": "Ouest",
-    "9": "Sud-Ouest",
-    "10": "Adamaoua",
-    "11": "Centre",
-    "12": "Extreme-Nord",
-    "13": "Nord",
-    "14": "Sud"
-  },
-  "CN": {
-    "1": "Anhui",
-    "2": "Zhejiang",
-    "3": "Jiangxi",
-    "4": "Jiangsu",
-    "5": "Jilin",
-    "6": "Qinghai",
-    "7": "Fujian",
-    "8": "Heilongjiang",
-    "9": "Henan",
-    "10": "Hebei",
-    "11": "Hunan",
-    "12": "Hubei",
-    "13": "Xinjiang",
-    "14": "Xizang",
-    "15": "Gansu",
-    "16": "Guangxi",
-    "18": "Guizhou",
-    "19": "Liaoning",
-    "20": "Nei Mongol",
-    "21": "Ningxia",
-    "22": "Beijing",
-    "23": "Shanghai",
-    "24": "Shanxi",
-    "25": "Shandong",
-    "26": "Shaanxi",
-    "28": "Tianjin",
-    "29": "Yunnan",
-    "30": "Guangdong",
-    "31": "Hainan",
-    "32": "Sichuan",
-    "33": "Chongqing"
-  },
-  "CO": {
-    "1": "Amazonas",
-    "2": "Antioquia",
-    "3": "Arauca",
-    "4": "Atlantico",
-    "8": "Caqueta",
-    "9": "Cauca",
-    "10": "Cesar",
-    "11": "Choco",
-    "12": "Cordoba",
-    "14": "Guaviare",
-    "15": "Guainia",
-    "16": "Huila",
-    "17": "La Guajira",
-    "19": "Meta",
-    "20": "Narino",
-    "21": "Norte de Santander",
-    "22": "Putumayo",
-    "23": "Quindio",
-    "24": "Risaralda",
-    "25": "San Andres y Providencia",
-    "26": "Santander",
-    "27": "Sucre",
-    "28": "Tolima",
-    "29": "Valle del Cauca",
-    "30": "Vaupes",
-    "31": "Vichada",
-    "32": "Casanare",
-    "33": "Cundinamarca",
-    "34": "Distrito Especial",
-    "35": "Bolivar",
-    "36": "Boyaca",
-    "37": "Caldas",
-    "38": "Magdalena"
-  },
-  "CR": {
-    "1": "Alajuela",
-    "2": "Cartago",
-    "3": "Guanacaste",
-    "4": "Heredia",
-    "6": "Limon",
-    "7": "Puntarenas",
-    "8": "San Jose"
-  },
-  "CU": {
-    "1": "Pinar del Rio",
-    "2": "Ciudad de la Habana",
-    "3": "Matanzas",
-    "4": "Isla de la Juventud",
-    "5": "Camaguey",
-    "7": "Ciego de Avila",
-    "8": "Cienfuegos",
-    "9": "Granma",
-    "10": "Guantanamo",
-    "11": "La Habana",
-    "12": "Holguin",
-    "13": "Las Tunas",
-    "14": "Sancti Spiritus",
-    "15": "Santiago de Cuba",
-    "16": "Villa Clara"
-  },
-  "CV": {
-    "1": "Boa Vista",
-    "2": "Brava",
-    "4": "Maio",
-    "5": "Paul",
-    "7": "Ribeira Grande",
-    "8": "Sal",
-    "10": "Sao Nicolau",
-    "11": "Sao Vicente",
-    "13": "Mosteiros",
-    "14": "Praia",
-    "15": "Santa Catarina",
-    "16": "Santa Cruz",
-    "17": "Sao Domingos",
-    "18": "Sao Filipe",
-    "19": "Sao Miguel",
-    "20": "Tarrafal"
-  },
-  "CY": {
-    "1": "Famagusta",
-    "2": "Kyrenia",
-    "3": "Larnaca",
-    "4": "Nicosia",
-    "5": "Limassol",
-    "6": "Paphos"
-  },
-  "CZ": {
-    "52": "Hlavni mesto Praha",
-    "78": "Jihomoravsky kraj",
-    "79": "Jihocesky kraj",
-    "80": "Vysocina",
-    "81": "Karlovarsky kraj",
-    "82": "Kralovehradecky kraj",
-    "83": "Liberecky kraj",
-    "84": "Olomoucky kraj",
-    "85": "Moravskoslezsky kraj",
-    "86": "Pardubicky kraj",
-    "87": "Plzensky kraj",
-    "88": "Stredocesky kraj",
-    "89": "Ustecky kraj",
-    "90": "Zlinsky kraj"
-  },
-  "DE": {
-    "1": "Baden-Wurttemberg",
-    "2": "Bayern",
-    "3": "Bremen",
-    "4": "Hamburg",
-    "5": "Hessen",
-    "6": "Niedersachsen",
-    "7": "Nordrhein-Westfalen",
-    "8": "Rheinland-Pfalz",
-    "9": "Saarland",
-    "10": "Schleswig-Holstein",
-    "11": "Brandenburg",
-    "12": "Mecklenburg-Vorpommern",
-    "13": "Sachsen",
-    "14": "Sachsen-Anhalt",
-    "15": "Thuringen",
-    "16": "Berlin"
-  },
-  "DJ": {
-    "1": "Ali Sabieh",
-    "4": "Obock",
-    "5": "Tadjoura",
-    "6": "Dikhil",
-    "7": "Djibouti",
-    "8": "Arta"
-  },
-  "DK": {
-    "17": "Hovedstaden",
-    "18": "Midtjylland",
-    "19": "Nordjylland",
-    "20": "Sjelland",
-    "21": "Syddanmark"
-  },
-  "DM": {
-    "2": "Saint Andrew",
-    "3": "Saint David",
-    "4": "Saint George",
-    "5": "Saint John",
-    "6": "Saint Joseph",
-    "7": "Saint Luke",
-    "8": "Saint Mark",
-    "9": "Saint Patrick",
-    "10": "Saint Paul",
-    "11": "Saint Peter"
-  },
-  "DO": {
-    "1": "Azua",
-    "2": "Baoruco",
-    "3": "Barahona",
-    "4": "Dajabon",
-    "5": "Distrito Nacional",
-    "6": "Duarte",
-    "8": "Espaillat",
-    "9": "Independencia",
-    "10": "La Altagracia",
-    "11": "Elias Pina",
-    "12": "La Romana",
-    "14": "Maria Trinidad Sanchez",
-    "15": "Monte Cristi",
-    "16": "Pedernales",
-    "17": "Peravia",
-    "18": "Puerto Plata",
-    "19": "Salcedo",
-    "20": "Samana",
-    "21": "Sanchez Ramirez",
-    "23": "San Juan",
-    "24": "San Pedro De Macoris",
-    "25": "Santiago",
-    "26": "Santiago Rodriguez",
-    "27": "Valverde",
-    "28": "El Seibo",
-    "29": "Hato Mayor",
-    "30": "La Vega",
-    "31": "Monsenor Nouel",
-    "32": "Monte Plata",
-    "33": "San Cristobal",
-    "34": "Distrito Nacional",
-    "35": "Peravia",
-    "36": "San Jose de Ocoa",
-    "37": "Santo Domingo"
-  },
-  "DZ": {
-    "1": "Alger",
-    "3": "Batna",
-    "4": "Constantine",
-    "6": "Medea",
-    "7": "Mostaganem",
-    "9": "Oran",
-    "10": "Saida",
-    "12": "Setif",
-    "13": "Tiaret",
-    "14": "Tizi Ouzou",
-    "15": "Tlemcen",
-    "18": "Bejaia",
-    "19": "Biskra",
-    "20": "Blida",
-    "21": "Bouira",
-    "22": "Djelfa",
-    "23": "Guelma",
-    "24": "Jijel",
-    "25": "Laghouat",
-    "26": "Mascara",
-    "27": "M'sila",
-    "29": "Oum el Bouaghi",
-    "30": "Sidi Bel Abbes",
-    "31": "Skikda",
-    "33": "Tebessa",
-    "34": "Adrar",
-    "35": "Ain Defla",
-    "36": "Ain Temouchent",
-    "37": "Annaba",
-    "38": "Bechar",
-    "39": "Bordj Bou Arreridj",
-    "40": "Boumerdes",
-    "41": "Chlef",
-    "42": "El Bayadh",
-    "43": "El Oued",
-    "44": "El Tarf",
-    "45": "Ghardaia",
-    "46": "Illizi",
-    "47": "Khenchela",
-    "48": "Mila",
-    "49": "Naama",
-    "50": "Ouargla",
-    "51": "Relizane",
-    "52": "Souk Ahras",
-    "53": "Tamanghasset",
-    "54": "Tindouf",
-    "55": "Tipaza",
-    "56": "Tissemsilt"
-  },
-  "EC": {
-    "1": "Galapagos",
-    "2": "Azuay",
-    "3": "Bolivar",
-    "4": "Canar",
-    "5": "Carchi",
-    "6": "Chimborazo",
-    "7": "Cotopaxi",
-    "8": "El Oro",
-    "9": "Esmeraldas",
-    "10": "Guayas",
-    "11": "Imbabura",
-    "12": "Loja",
-    "13": "Los Rios",
-    "14": "Manabi",
-    "15": "Morona-Santiago",
-    "17": "Pastaza",
-    "18": "Pichincha",
-    "19": "Tungurahua",
-    "20": "Zamora-Chinchipe",
-    "22": "Sucumbios",
-    "23": "Napo",
-    "24": "Orellana"
-  },
-  "EE": {
-    "1": "Harjumaa",
-    "2": "Hiiumaa",
-    "3": "Ida-Virumaa",
-    "4": "Jarvamaa",
-    "5": "Jogevamaa",
-    "6": "Kohtla-Jarve",
-    "7": "Laanemaa",
-    "8": "Laane-Virumaa",
-    "9": "Narva",
-    "10": "Parnu",
-    "11": "Parnumaa",
-    "12": "Polvamaa",
-    "13": "Raplamaa",
-    "14": "Saaremaa",
-    "15": "Sillamae",
-    "16": "Tallinn",
-    "17": "Tartu",
-    "18": "Tartumaa",
-    "19": "Valgamaa",
-    "20": "Viljandimaa",
-    "21": "Vorumaa"
-  },
-  "EG": {
-    "1": "Ad Daqahliyah",
-    "2": "Al Bahr al Ahmar",
-    "3": "Al Buhayrah",
-    "4": "Al Fayyum",
-    "5": "Al Gharbiyah",
-    "6": "Al Iskandariyah",
-    "7": "Al Isma'iliyah",
-    "8": "Al Jizah",
-    "9": "Al Minufiyah",
-    "10": "Al Minya",
-    "11": "Al Qahirah",
-    "12": "Al Qalyubiyah",
-    "13": "Al Wadi al Jadid",
-    "14": "Ash Sharqiyah",
-    "15": "As Suways",
-    "16": "Aswan",
-    "17": "Asyut",
-    "18": "Bani Suwayf",
-    "19": "Bur Sa'id",
-    "20": "Dumyat",
-    "21": "Kafr ash Shaykh",
-    "22": "Matruh",
-    "23": "Qina",
-    "24": "Suhaj",
-    "26": "Janub Sina'",
-    "27": "Shamal Sina'",
-    "28": "Al Uqsur"
-  },
-  "ER": {
-    "1": "Anseba",
-    "2": "Debub",
-    "3": "Debubawi K'eyih Bahri",
-    "4": "Gash Barka",
-    "5": "Ma'akel",
-    "6": "Semenawi K'eyih Bahri"
-  },
-  "ES": {
-    "7": "Islas Baleares",
-    "27": "La Rioja",
-    "29": "Madrid",
-    "31": "Murcia",
-    "32": "Navarra",
-    "34": "Asturias",
-    "39": "Cantabria",
-    "51": "Andalucia",
-    "52": "Aragon",
-    "53": "Canarias",
-    "54": "Castilla-La Mancha",
-    "55": "Castilla y Leon",
-    "56": "Catalonia",
-    "57": "Extremadura",
-    "58": "Galicia",
-    "59": "Pais Vasco",
-    "60": "Comunidad Valenciana"
-  },
-  "ET": {
-    "44": "Adis Abeba",
-    "45": "Afar",
-    "46": "Amara",
-    "47": "Binshangul Gumuz",
-    "48": "Dire Dawa",
-    "49": "Gambela Hizboch",
-    "50": "Hareri Hizb",
-    "51": "Oromiya",
-    "52": "Sumale",
-    "53": "Tigray",
-    "54": "YeDebub Biheroch Bihereseboch na Hizboch"
-  },
-  "FI": {
-    "1": "Aland",
-    "6": "Lapland",
-    "8": "Oulu",
-    "13": "Southern Finland",
-    "14": "Eastern Finland",
-    "15": "Western Finland"
-  },
-  "FJ": {
-    "1": "Central",
-    "2": "Eastern",
-    "3": "Northern",
-    "4": "Rotuma",
-    "5": "Western"
-  },
-  "FM": {
-    "1": "Kosrae",
-    "2": "Pohnpei",
-    "3": "Chuuk",
-    "4": "Yap"
-  },
-  "FR": {
-    "97": "Aquitaine",
-    "98": "Auvergne",
-    "99": "Basse-Normandie",
-    "832": "Bourgogne",
-    "833": "Bretagne",
-    "834": "Centre",
-    "835": "Champagne-Ardenne",
-    "836": "Corse",
-    "837": "Franche-Comte",
-    "838": "Haute-Normandie",
-    "839": "Ile-de-France",
-    "840": "Languedoc-Roussillon",
-    "875": "Limousin",
-    "876": "Lorraine",
-    "877": "Midi-Pyrenees",
-    "878": "Nord-Pas-de-Calais",
-    "879": "Pays de la Loire",
-    "880": "Picardie",
-    "881": "Poitou-Charentes",
-    "882": "Provence-Alpes-Cote d'Azur",
-    "883": "Rhone-Alpes",
-    "918": "Alsace"
-  },
-  "GA": {
-    "1": "Estuaire",
-    "2": "Haut-Ogooue",
-    "3": "Moyen-Ogooue",
-    "4": "Ngounie",
-    "5": "Nyanga",
-    "6": "Ogooue-Ivindo",
-    "7": "Ogooue-Lolo",
-    "8": "Ogooue-Maritime",
-    "9": "Woleu-Ntem"
-  },
-  "GB": {
-    "832": "Barking and Dagenham",
-    "833": "Barnet",
-    "834": "Barnsley",
-    "835": "Bath and North East Somerset",
-    "836": "Bedfordshire",
-    "837": "Bexley",
-    "838": "Birmingham",
-    "839": "Blackburn with Darwen",
-    "840": "Blackpool",
-    "875": "Bolton",
-    "876": "Bournemouth",
-    "877": "Bracknell Forest",
-    "878": "Bradford",
-    "879": "Brent",
-    "880": "Brighton and Hove",
-    "881": "Bristol",
-    "882": "Bromley",
-    "883": "Buckinghamshire",
-    "918": "Bury",
-    "919": "Calderdale",
-    "920": "Cambridgeshire",
-    "921": "Camden",
-    "922": "Cheshire",
-    "923": "Cornwall",
-    "924": "Coventry",
-    "925": "Croydon",
-    "926": "Cumbria",
-    "961": "Darlington",
-    "962": "Derby",
-    "963": "Derbyshire",
-    "964": "Devon",
-    "965": "Doncaster",
-    "966": "Dorset",
-    "967": "Dudley",
-    "968": "Durham",
-    "969": "Ealing",
-    "1004": "East Riding of Yorkshire",
-    "1005": "East Sussex",
-    "1006": "Enfield",
-    "1007": "Essex",
-    "1008": "Gateshead",
-    "1009": "Gloucestershire",
-    "1010": "Greenwich",
-    "1011": "Hackney",
-    "1012": "Halton",
-    "1047": "Hammersmith and Fulham",
-    "1048": "Hampshire",
-    "1049": "Haringey",
-    "1050": "Harrow",
-    "1051": "Hartlepool",
-    "1052": "Havering",
-    "1053": "Herefordshire",
-    "1054": "Hertford",
-    "1055": "Hillingdon",
-    "1090": "Hounslow",
-    "1091": "Isle of Wight",
-    "1092": "Islington",
-    "1093": "Kensington and Chelsea",
-    "1094": "Kent",
-    "1095": "Kingston upon Hull",
-    "1096": "Kingston upon Thames",
-    "1097": "Kirklees",
-    "1098": "Knowsley",
-    "1133": "Lambeth",
-    "1134": "Lancashire",
-    "1135": "Leeds",
-    "1136": "Leicester",
-    "1137": "Leicestershire",
-    "1138": "Lewisham",
-    "1139": "Lincolnshire",
-    "1140": "Liverpool",
-    "1141": "London",
-    "1176": "Luton",
-    "1177": "Manchester",
-    "1178": "Medway",
-    "1179": "Merton",
-    "1180": "Middlesbrough",
-    "1181": "Milton Keynes",
-    "1182": "Newcastle upon Tyne",
-    "1183": "Newham",
-    "1184": "Norfolk",
-    "1219": "Northamptonshire",
-    "1220": "North East Lincolnshire",
-    "1221": "North Lincolnshire",
-    "1222": "North Somerset",
-    "1223": "North Tyneside",
-    "1224": "Northumberland",
-    "1225": "North Yorkshire",
-    "1226": "Nottingham",
-    "1227": "Nottinghamshire",
-    "1262": "Oldham",
-    "1263": "Oxfordshire",
-    "1264": "Peterborough",
-    "1265": "Plymouth",
-    "1266": "Poole",
-    "1267": "Portsmouth",
-    "1268": "Reading",
-    "1269": "Redbridge",
-    "1270": "Redcar and Cleveland",
-    "1305": "Richmond upon Thames",
-    "1306": "Rochdale",
-    "1307": "Rotherham",
-    "1308": "Rutland",
-    "1309": "Salford",
-    "1310": "Shropshire",
-    "1311": "Sandwell",
-    "1312": "Sefton",
-    "1313": "Sheffield",
-    "1348": "Slough",
-    "1349": "Solihull",
-    "1350": "Somerset",
-    "1351": "Southampton",
-    "1352": "Southend-on-Sea",
-    "1353": "South Gloucestershire",
-    "1354": "South Tyneside",
-    "1355": "Southwark",
-    "1356": "Staffordshire",
-    "1391": "St. Helens",
-    "1392": "Stockport",
-    "1393": "Stockton-on-Tees",
-    "1394": "Stoke-on-Trent",
-    "1395": "Suffolk",
-    "1396": "Sunderland",
-    "1397": "Surrey",
-    "1398": "Sutton",
-    "1399": "Swindon",
-    "1434": "Tameside",
-    "1435": "Telford and Wrekin",
-    "1436": "Thurrock",
-    "1437": "Torbay",
-    "1438": "Tower Hamlets",
-    "1439": "Trafford",
-    "1440": "Wakefield",
-    "1441": "Walsall",
-    "1442": "Waltham Forest",
-    "1477": "Wandsworth",
-    "1478": "Warrington",
-    "1479": "Warwickshire",
-    "1480": "West Berkshire",
-    "1481": "Westminster",
-    "1482": "West Sussex",
-    "1483": "Wigan",
-    "1484": "Wiltshire",
-    "1485": "Windsor and Maidenhead",
-    "1520": "Wirral",
-    "1521": "Wokingham",
-    "1522": "Wolverhampton",
-    "1523": "Worcestershire",
-    "1524": "York",
-    "1525": "Antrim",
-    "1526": "Ards",
-    "1527": "Armagh",
-    "1528": "Ballymena",
-    "1563": "Ballymoney",
-    "1564": "Banbridge",
-    "1565": "Belfast",
-    "1566": "Carrickfergus",
-    "1567": "Castlereagh",
-    "1568": "Coleraine",
-    "1569": "Cookstown",
-    "1570": "Craigavon",
-    "1571": "Down",
-    "1606": "Dungannon",
-    "1607": "Fermanagh",
-    "1608": "Larne",
-    "1609": "Limavady",
-    "1610": "Lisburn",
-    "1611": "Derry",
-    "1612": "Magherafelt",
-    "1613": "Moyle",
-    "1614": "Newry and Mourne",
-    "1649": "Newtownabbey",
-    "1650": "North Down",
-    "1651": "Omagh",
-    "1652": "Strabane",
-    "1653": "Aberdeen City",
-    "1654": "Aberdeenshire",
-    "1655": "Angus",
-    "1656": "Argyll and Bute",
-    "1657": "Scottish Borders",
-    "1692": "Clackmannanshire",
-    "1693": "Dumfries and Galloway",
-    "1694": "Dundee City",
-    "1695": "East Ayrshire",
-    "1696": "East Dunbartonshire",
-    "1697": "East Lothian",
-    "1698": "East Renfrewshire",
-    "1699": "Edinburgh",
-    "1700": "Falkirk",
-    "1735": "Fife",
-    "1736": "Glasgow City",
-    "1737": "Highland",
-    "1738": "Inverclyde",
-    "1739": "Midlothian",
-    "1740": "Moray",
-    "1741": "North Ayrshire",
-    "1742": "North Lanarkshire",
-    "1743": "Orkney",
-    "1778": "Perth and Kinross",
-    "1779": "Renfrewshire",
-    "1780": "Shetland Islands",
-    "1781": "South Ayrshire",
-    "1782": "South Lanarkshire",
-    "1783": "Stirling",
-    "1784": "West Dunbartonshire",
-    "1785": "Eilean Siar",
-    "1786": "West Lothian",
-    "1821": "Isle of Anglesey",
-    "1822": "Blaenau Gwent",
-    "1823": "Bridgend",
-    "1824": "Caerphilly",
-    "1825": "Cardiff",
-    "1826": "Ceredigion",
-    "1827": "Carmarthenshire",
-    "1828": "Conwy",
-    "1829": "Denbighshire",
-    "1864": "Flintshire",
-    "1865": "Gwynedd",
-    "1866": "Merthyr Tydfil",
-    "1867": "Monmouthshire",
-    "1868": "Neath Port Talbot",
-    "1869": "Newport",
-    "1870": "Pembrokeshire",
-    "1871": "Powys",
-    "1872": "Rhondda Cynon Taff",
-    "1907": "Swansea",
-    "1908": "Torfaen",
-    "1909": "Vale of Glamorgan",
-    "1910": "Wrexham",
-    "1911": "Bedfordshire",
-    "1912": "Central Bedfordshire",
-    "1913": "Cheshire East",
-    "1914": "Cheshire West and Chester",
-    "1915": "Isles of Scilly"
-  },
-  "GD": {
-    "1": "Saint Andrew",
-    "2": "Saint David",
-    "3": "Saint George",
-    "4": "Saint John",
-    "5": "Saint Mark",
-    "6": "Saint Patrick"
-  },
-  "GE": {
-    "1": "Abashis Raioni",
-    "2": "Abkhazia",
-    "3": "Adigenis Raioni",
-    "4": "Ajaria",
-    "5": "Akhalgoris Raioni",
-    "6": "Akhalk'alak'is Raioni",
-    "7": "Akhalts'ikhis Raioni",
-    "8": "Akhmetis Raioni",
-    "9": "Ambrolauris Raioni",
-    "10": "Aspindzis Raioni",
-    "11": "Baghdat'is Raioni",
-    "12": "Bolnisis Raioni",
-    "13": "Borjomis Raioni",
-    "14": "Chiat'ura",
-    "15": "Ch'khorotsqus Raioni",
-    "16": "Ch'okhatauris Raioni",
-    "17": "Dedop'listsqaros Raioni",
-    "18": "Dmanisis Raioni",
-    "19": "Dushet'is Raioni",
-    "20": "Gardabanis Raioni",
-    "21": "Gori",
-    "22": "Goris Raioni",
-    "23": "Gurjaanis Raioni",
-    "24": "Javis Raioni",
-    "25": "K'arelis Raioni",
-    "26": "Kaspis Raioni",
-    "27": "Kharagaulis Raioni",
-    "28": "Khashuris Raioni",
-    "29": "Khobis Raioni",
-    "30": "Khonis Raioni",
-    "31": "K'ut'aisi",
-    "32": "Lagodekhis Raioni",
-    "33": "Lanch'khut'is Raioni",
-    "34": "Lentekhis Raioni",
-    "35": "Marneulis Raioni",
-    "36": "Martvilis Raioni",
-    "37": "Mestiis Raioni",
-    "38": "Mts'khet'is Raioni",
-    "39": "Ninotsmindis Raioni",
-    "40": "Onis Raioni",
-    "41": "Ozurget'is Raioni",
-    "42": "P'ot'i",
-    "43": "Qazbegis Raioni",
-    "44": "Qvarlis Raioni",
-    "45": "Rust'avi",
-    "46": "Sach'kheris Raioni",
-    "47": "Sagarejos Raioni",
-    "48": "Samtrediis Raioni",
-    "49": "Senakis Raioni",
-    "50": "Sighnaghis Raioni",
-    "51": "T'bilisi",
-    "52": "T'elavis Raioni",
-    "53": "T'erjolis Raioni",
-    "54": "T'et'ritsqaros Raioni",
-    "55": "T'ianet'is Raioni",
-    "56": "Tqibuli",
-    "57": "Ts'ageris Raioni",
-    "58": "Tsalenjikhis Raioni",
-    "59": "Tsalkis Raioni",
-    "60": "Tsqaltubo",
-    "61": "Vanis Raioni",
-    "62": "Zestap'onis Raioni",
-    "63": "Zugdidi",
-    "64": "Zugdidis Raioni"
-  },
-  "GH": {
-    "1": "Greater Accra",
-    "2": "Ashanti",
-    "3": "Brong-Ahafo",
-    "4": "Central",
-    "5": "Eastern",
-    "6": "Northern",
-    "8": "Volta",
-    "9": "Western",
-    "10": "Upper East",
-    "11": "Upper West"
-  },
-  "GL": {
-    "1": "Nordgronland",
-    "2": "Ostgronland",
-    "3": "Vestgronland"
-  },
-  "GM": {
-    "1": "Banjul",
-    "2": "Lower River",
-    "3": "Central River",
-    "4": "Upper River",
-    "5": "Western",
-    "7": "North Bank"
-  },
-  "GN": {
-    "1": "Beyla",
-    "2": "Boffa",
-    "3": "Boke",
-    "4": "Conakry",
-    "5": "Dabola",
-    "6": "Dalaba",
-    "7": "Dinguiraye",
-    "9": "Faranah",
-    "10": "Forecariah",
-    "11": "Fria",
-    "12": "Gaoual",
-    "13": "Gueckedou",
-    "15": "Kerouane",
-    "16": "Kindia",
-    "17": "Kissidougou",
-    "18": "Koundara",
-    "19": "Kouroussa",
-    "21": "Macenta",
-    "22": "Mali",
-    "23": "Mamou",
-    "25": "Pita",
-    "27": "Telimele",
-    "28": "Tougue",
-    "29": "Yomou",
-    "30": "Coyah",
-    "31": "Dubreka",
-    "32": "Kankan",
-    "33": "Koubia",
-    "34": "Labe",
-    "35": "Lelouma",
-    "36": "Lola",
-    "37": "Mandiana",
-    "38": "Nzerekore",
-    "39": "Siguiri"
-  },
-  "GQ": {
-    "3": "Annobon",
-    "4": "Bioko Norte",
-    "5": "Bioko Sur",
-    "6": "Centro Sur",
-    "7": "Kie-Ntem",
-    "8": "Litoral",
-    "9": "Wele-Nzas"
-  },
-  "GR": {
-    "1": "Evros",
-    "2": "Rodhopi",
-    "3": "Xanthi",
-    "4": "Drama",
-    "5": "Serrai",
-    "6": "Kilkis",
-    "7": "Pella",
-    "8": "Florina",
-    "9": "Kastoria",
-    "10": "Grevena",
-    "11": "Kozani",
-    "12": "Imathia",
-    "13": "Thessaloniki",
-    "14": "Kavala",
-    "15": "Khalkidhiki",
-    "16": "Pieria",
-    "17": "Ioannina",
-    "18": "Thesprotia",
-    "19": "Preveza",
-    "20": "Arta",
-    "21": "Larisa",
-    "22": "Trikala",
-    "23": "Kardhitsa",
-    "24": "Magnisia",
-    "25": "Kerkira",
-    "26": "Levkas",
-    "27": "Kefallinia",
-    "28": "Zakinthos",
-    "29": "Fthiotis",
-    "30": "Evritania",
-    "31": "Aitolia kai Akarnania",
-    "32": "Fokis",
-    "33": "Voiotia",
-    "34": "Evvoia",
-    "35": "Attiki",
-    "36": "Argolis",
-    "37": "Korinthia",
-    "38": "Akhaia",
-    "39": "Ilia",
-    "40": "Messinia",
-    "41": "Arkadhia",
-    "42": "Lakonia",
-    "43": "Khania",
-    "44": "Rethimni",
-    "45": "Iraklion",
-    "46": "Lasithi",
-    "47": "Dhodhekanisos",
-    "48": "Samos",
-    "49": "Kikladhes",
-    "50": "Khios",
-    "51": "Lesvos"
-  },
-  "GT": {
-    "1": "Alta Verapaz",
-    "2": "Baja Verapaz",
-    "3": "Chimaltenango",
-    "4": "Chiquimula",
-    "5": "El Progreso",
-    "6": "Escuintla",
-    "7": "Guatemala",
-    "8": "Huehuetenango",
-    "9": "Izabal",
-    "10": "Jalapa",
-    "11": "Jutiapa",
-    "12": "Peten",
-    "13": "Quetzaltenango",
-    "14": "Quiche",
-    "15": "Retalhuleu",
-    "16": "Sacatepequez",
-    "17": "San Marcos",
-    "18": "Santa Rosa",
-    "19": "Solola",
-    "20": "Suchitepequez",
-    "21": "Totonicapan",
-    "22": "Zacapa"
-  },
-  "GW": {
-    "1": "Bafata",
-    "2": "Quinara",
-    "4": "Oio",
-    "5": "Bolama",
-    "6": "Cacheu",
-    "7": "Tombali",
-    "10": "Gabu",
-    "11": "Bissau",
-    "12": "Biombo"
-  },
-  "GY": {
-    "10": "Barima-Waini",
-    "11": "Cuyuni-Mazaruni",
-    "12": "Demerara-Mahaica",
-    "13": "East Berbice-Corentyne",
-    "14": "Essequibo Islands-West Demerara",
-    "15": "Mahaica-Berbice",
-    "16": "Pomeroon-Supenaam",
-    "17": "Potaro-Siparuni",
-    "18": "Upper Demerara-Berbice",
-    "19": "Upper Takutu-Upper Essequibo"
-  },
-  "HN": {
-    "1": "Atlantida",
-    "2": "Choluteca",
-    "3": "Colon",
-    "4": "Comayagua",
-    "5": "Copan",
-    "6": "Cortes",
-    "7": "El Paraiso",
-    "8": "Francisco Morazan",
-    "9": "Gracias a Dios",
-    "10": "Intibuca",
-    "11": "Islas de la Bahia",
-    "12": "La Paz",
-    "13": "Lempira",
-    "14": "Ocotepeque",
-    "15": "Olancho",
-    "16": "Santa Barbara",
-    "17": "Valle",
-    "18": "Yoro"
-  },
-  "HR": {
-    "1": "Bjelovarsko-Bilogorska",
-    "2": "Brodsko-Posavska",
-    "3": "Dubrovacko-Neretvanska",
-    "4": "Istarska",
-    "5": "Karlovacka",
-    "6": "Koprivnicko-Krizevacka",
-    "7": "Krapinsko-Zagorska",
-    "8": "Licko-Senjska",
-    "9": "Medimurska",
-    "10": "Osjecko-Baranjska",
-    "11": "Pozesko-Slavonska",
-    "12": "Primorsko-Goranska",
-    "13": "Sibensko-Kninska",
-    "14": "Sisacko-Moslavacka",
-    "15": "Splitsko-Dalmatinska",
-    "16": "Varazdinska",
-    "17": "Viroviticko-Podravska",
-    "18": "Vukovarsko-Srijemska",
-    "19": "Zadarska",
-    "20": "Zagrebacka",
-    "21": "Grad Zagreb"
-  },
-  "HT": {
-    "3": "Nord-Ouest",
-    "6": "Artibonite",
-    "7": "Centre",
-    "9": "Nord",
-    "10": "Nord-Est",
-    "11": "Ouest",
-    "12": "Sud",
-    "13": "Sud-Est",
-    "14": "Grand' Anse",
-    "15": "Nippes"
-  },
-  "HU": {
-    "1": "Bacs-Kiskun",
-    "2": "Baranya",
-    "3": "Bekes",
-    "4": "Borsod-Abauj-Zemplen",
-    "5": "Budapest",
-    "6": "Csongrad",
-    "7": "Debrecen",
-    "8": "Fejer",
-    "9": "Gyor-Moson-Sopron",
-    "10": "Hajdu-Bihar",
-    "11": "Heves",
-    "12": "Komarom-Esztergom",
-    "13": "Miskolc",
-    "14": "Nograd",
-    "15": "Pecs",
-    "16": "Pest",
-    "17": "Somogy",
-    "18": "Szabolcs-Szatmar-Bereg",
-    "19": "Szeged",
-    "20": "Jasz-Nagykun-Szolnok",
-    "21": "Tolna",
-    "22": "Vas",
-    "23": "Veszprem",
-    "24": "Zala",
-    "25": "Gyor",
-    "26": "Bekescsaba",
-    "27": "Dunaujvaros",
-    "28": "Eger",
-    "29": "Hodmezovasarhely",
-    "30": "Kaposvar",
-    "31": "Kecskemet",
-    "32": "Nagykanizsa",
-    "33": "Nyiregyhaza",
-    "34": "Sopron",
-    "35": "Szekesfehervar",
-    "36": "Szolnok",
-    "37": "Szombathely",
-    "38": "Tatabanya",
-    "39": "Veszprem",
-    "40": "Zalaegerszeg",
-    "41": "Salgotarjan",
-    "42": "Szekszard",
-    "43": "Erd"
-  },
-  "ID": {
-    "1": "Aceh",
-    "2": "Bali",
-    "3": "Bengkulu",
-    "4": "Jakarta Raya",
-    "5": "Jambi",
-    "7": "Jawa Tengah",
-    "8": "Jawa Timur",
-    "10": "Yogyakarta",
-    "11": "Kalimantan Barat",
-    "12": "Kalimantan Selatan",
-    "13": "Kalimantan Tengah",
-    "14": "Kalimantan Timur",
-    "15": "Lampung",
-    "17": "Nusa Tenggara Barat",
-    "18": "Nusa Tenggara Timur",
-    "21": "Sulawesi Tengah",
-    "22": "Sulawesi Tenggara",
-    "24": "Sumatera Barat",
-    "26": "Sumatera Utara",
-    "28": "Maluku",
-    "29": "Maluku Utara",
-    "30": "Jawa Barat",
-    "31": "Sulawesi Utara",
-    "32": "Sumatera Selatan",
-    "33": "Banten",
-    "34": "Gorontalo",
-    "35": "Kepulauan Bangka Belitung",
-    "36": "Papua",
-    "37": "Riau",
-    "38": "Sulawesi Selatan",
-    "39": "Irian Jaya Barat",
-    "40": "Kepulauan Riau",
-    "41": "Sulawesi Barat"
-  },
-  "IE": {
-    "1": "Carlow",
-    "2": "Cavan",
-    "3": "Clare",
-    "4": "Cork",
-    "6": "Donegal",
-    "7": "Dublin",
-    "10": "Galway",
-    "11": "Kerry",
-    "12": "Kildare",
-    "13": "Kilkenny",
-    "14": "Leitrim",
-    "15": "Laois",
-    "16": "Limerick",
-    "18": "Longford",
-    "19": "Louth",
-    "20": "Mayo",
-    "21": "Meath",
-    "22": "Monaghan",
-    "23": "Offaly",
-    "24": "Roscommon",
-    "25": "Sligo",
-    "26": "Tipperary",
-    "27": "Waterford",
-    "29": "Westmeath",
-    "30": "Wexford",
-    "31": "Wicklow"
-  },
-  "IL": {
-    "1": "HaDarom",
-    "2": "HaMerkaz",
-    "3": "HaZafon",
-    "4": "Hefa",
-    "5": "Tel Aviv",
-    "6": "Yerushalayim"
-  },
-  "IN": {
-    "1": "Andaman and Nicobar Islands",
-    "2": "Andhra Pradesh",
-    "3": "Assam",
-    "5": "Chandigarh",
-    "6": "Dadra and Nagar Haveli",
-    "7": "Delhi",
-    "9": "Gujarat",
-    "10": "Haryana",
-    "11": "Himachal Pradesh",
-    "12": "Jammu and Kashmir",
-    "13": "Kerala",
-    "14": "Lakshadweep",
-    "16": "Maharashtra",
-    "17": "Manipur",
-    "18": "Meghalaya",
-    "19": "Karnataka",
-    "20": "Nagaland",
-    "21": "Orissa",
-    "22": "Puducherry",
-    "23": "Punjab",
-    "24": "Rajasthan",
-    "25": "Tamil Nadu",
-    "26": "Tripura",
-    "28": "West Bengal",
-    "29": "Sikkim",
-    "30": "Arunachal Pradesh",
-    "31": "Mizoram",
-    "32": "Daman and Diu",
-    "33": "Goa",
-    "34": "Bihar",
-    "35": "Madhya Pradesh",
-    "36": "Uttar Pradesh",
-    "37": "Chhattisgarh",
-    "38": "Jharkhand",
-    "39": "Uttarakhand"
-  },
-  "IQ": {
-    "1": "Al Anbar",
-    "2": "Al Basrah",
-    "3": "Al Muthanna",
-    "4": "Al Qadisiyah",
-    "5": "As Sulaymaniyah",
-    "6": "Babil",
-    "7": "Baghdad",
-    "8": "Dahuk",
-    "9": "Dhi Qar",
-    "10": "Diyala",
-    "11": "Arbil",
-    "12": "Karbala'",
-    "13": "At Ta'mim",
-    "14": "Maysan",
-    "15": "Ninawa",
-    "16": "Wasit",
-    "17": "An Najaf",
-    "18": "Salah ad Din"
-  },
-  "IR": {
-    "1": "Azarbayjan-e Bakhtari",
-    "3": "Chahar Mahall va Bakhtiari",
-    "4": "Sistan va Baluchestan",
-    "5": "Kohkiluyeh va Buyer Ahmadi",
-    "7": "Fars",
-    "8": "Gilan",
-    "9": "Hamadan",
-    "10": "Ilam",
-    "11": "Hormozgan",
-    "12": "Kerman",
-    "13": "Bakhtaran",
-    "15": "Khuzestan",
-    "16": "Kordestan",
-    "17": "Mazandaran",
-    "18": "Semnan Province",
-    "19": "Markazi",
-    "21": "Zanjan",
-    "22": "Bushehr",
-    "23": "Lorestan",
-    "24": "Markazi",
-    "25": "Semnan",
-    "26": "Tehran",
-    "27": "Zanjan",
-    "28": "Esfahan",
-    "29": "Kerman",
-    "30": "Khorasan",
-    "31": "Yazd",
-    "32": "Ardabil",
-    "33": "East Azarbaijan",
-    "34": "Markazi",
-    "35": "Mazandaran",
-    "36": "Zanjan",
-    "37": "Golestan",
-    "38": "Qazvin",
-    "39": "Qom",
-    "40": "Yazd",
-    "41": "Khorasan-e Janubi",
-    "42": "Khorasan-e Razavi",
-    "43": "Khorasan-e Shemali",
-    "44": "Alborz"
-  },
-  "IS": {
-    "3": "Arnessysla",
-    "5": "Austur-Hunavatnssysla",
-    "6": "Austur-Skaftafellssysla",
-    "7": "Borgarfjardarsysla",
-    "9": "Eyjafjardarsysla",
-    "10": "Gullbringusysla",
-    "15": "Kjosarsysla",
-    "17": "Myrasysla",
-    "20": "Nordur-Mulasysla",
-    "21": "Nordur-Tingeyjarsysla",
-    "23": "Rangarvallasysla",
-    "28": "Skagafjardarsysla",
-    "29": "Snafellsnes- og Hnappadalssysla",
-    "30": "Strandasysla",
-    "31": "Sudur-Mulasysla",
-    "32": "Sudur-Tingeyjarsysla",
-    "34": "Vestur-Bardastrandarsysla",
-    "35": "Vestur-Hunavatnssysla",
-    "36": "Vestur-Isafjardarsysla",
-    "37": "Vestur-Skaftafellssysla",
-    "38": "Austurland",
-    "39": "Hofuoborgarsvaoio",
-    "40": "Norourland Eystra",
-    "41": "Norourland Vestra",
-    "42": "Suourland",
-    "43": "Suournes",
-    "44": "Vestfiroir",
-    "45": "Vesturland"
-  },
-  "IT": {
-    "1": "Abruzzi",
-    "2": "Basilicata",
-    "3": "Calabria",
-    "4": "Campania",
-    "5": "Emilia-Romagna",
-    "6": "Friuli-Venezia Giulia",
-    "7": "Lazio",
-    "8": "Liguria",
-    "9": "Lombardia",
-    "10": "Marche",
-    "11": "Molise",
-    "12": "Piemonte",
-    "13": "Puglia",
-    "14": "Sardegna",
-    "15": "Sicilia",
-    "16": "Toscana",
-    "17": "Trentino-Alto Adige",
-    "18": "Umbria",
-    "19": "Valle d'Aosta",
-    "20": "Veneto"
-  },
-  "JM": {
-    "1": "Clarendon",
-    "2": "Hanover",
-    "4": "Manchester",
-    "7": "Portland",
-    "8": "Saint Andrew",
-    "9": "Saint Ann",
-    "10": "Saint Catherine",
-    "11": "Saint Elizabeth",
-    "12": "Saint James",
-    "13": "Saint Mary",
-    "14": "Saint Thomas",
-    "15": "Trelawny",
-    "16": "Westmoreland",
-    "17": "Kingston"
-  },
-  "JO": {
-    "2": "Al Balqa'",
-    "9": "Al Karak",
-    "12": "At Tafilah",
-    "15": "Al Mafraq",
-    "16": "Amman",
-    "17": "Az Zaraqa",
-    "18": "Irbid",
-    "19": "Ma'an",
-    "20": "Ajlun",
-    "21": "Al Aqabah",
-    "22": "Jarash",
-    "23": "Madaba"
-  },
-  "JP": {
-    "1": "Aichi",
-    "2": "Akita",
-    "3": "Aomori",
-    "4": "Chiba",
-    "5": "Ehime",
-    "6": "Fukui",
-    "7": "Fukuoka",
-    "8": "Fukushima",
-    "9": "Gifu",
-    "10": "Gumma",
-    "11": "Hiroshima",
-    "12": "Hokkaido",
-    "13": "Hyogo",
-    "14": "Ibaraki",
-    "15": "Ishikawa",
-    "16": "Iwate",
-    "17": "Kagawa",
-    "18": "Kagoshima",
-    "19": "Kanagawa",
-    "20": "Kochi",
-    "21": "Kumamoto",
-    "22": "Kyoto",
-    "23": "Mie",
-    "24": "Miyagi",
-    "25": "Miyazaki",
-    "26": "Nagano",
-    "27": "Nagasaki",
-    "28": "Nara",
-    "29": "Niigata",
-    "30": "Oita",
-    "31": "Okayama",
-    "32": "Osaka",
-    "33": "Saga",
-    "34": "Saitama",
-    "35": "Shiga",
-    "36": "Shimane",
-    "37": "Shizuoka",
-    "38": "Tochigi",
-    "39": "Tokushima",
-    "40": "Tokyo",
-    "41": "Tottori",
-    "42": "Toyama",
-    "43": "Wakayama",
-    "44": "Yamagata",
-    "45": "Yamaguchi",
-    "46": "Yamanashi",
-    "47": "Okinawa"
-  },
-  "KE": {
-    "1": "Central",
-    "2": "Coast",
-    "3": "Eastern",
-    "5": "Nairobi Area",
-    "6": "North-Eastern",
-    "7": "Nyanza",
-    "8": "Rift Valley",
-    "9": "Western"
-  },
-  "KG": {
-    "1": "Bishkek",
-    "2": "Chuy",
-    "3": "Jalal-Abad",
-    "4": "Naryn",
-    "5": "Osh",
-    "6": "Talas",
-    "7": "Ysyk-Kol",
-    "8": "Osh",
-    "9": "Batken"
-  },
-  "KH": {
-    "1": "Batdambang",
-    "2": "Kampong Cham",
-    "3": "Kampong Chhnang",
-    "4": "Kampong Speu",
-    "5": "Kampong Thum",
-    "6": "Kampot",
-    "7": "Kandal",
-    "8": "Koh Kong",
-    "9": "Kracheh",
-    "10": "Mondulkiri",
-    "11": "Phnum Penh",
-    "12": "Pursat",
-    "13": "Preah Vihear",
-    "14": "Prey Veng",
-    "15": "Ratanakiri Kiri",
-    "16": "Siem Reap",
-    "17": "Stung Treng",
-    "18": "Svay Rieng",
-    "19": "Takeo",
-    "22": "Phnum Penh",
-    "23": "Ratanakiri",
-    "25": "Banteay Meanchey",
-    "28": "Preah Seihanu",
-    "29": "Batdambang",
-    "30": "Pailin"
-  },
-  "KI": {
-    "1": "Gilbert Islands",
-    "2": "Line Islands",
-    "3": "Phoenix Islands"
-  },
-  "KM": {
-    "1": "Anjouan",
-    "2": "Grande Comore",
-    "3": "Moheli"
-  },
-  "KN": {
-    "1": "Christ Church Nichola Town",
-    "2": "Saint Anne Sandy Point",
-    "3": "Saint George Basseterre",
-    "4": "Saint George Gingerland",
-    "5": "Saint James Windward",
-    "6": "Saint John Capisterre",
-    "7": "Saint John Figtree",
-    "8": "Saint Mary Cayon",
-    "9": "Saint Paul Capisterre",
-    "10": "Saint Paul Charlestown",
-    "11": "Saint Peter Basseterre",
-    "12": "Saint Thomas Lowland",
-    "13": "Saint Thomas Middle Island",
-    "15": "Trinity Palmetto Point"
-  },
-  "KP": {
-    "1": "Chagang-do",
-    "3": "Hamgyong-namdo",
-    "6": "Hwanghae-namdo",
-    "7": "Hwanghae-bukto",
-    "8": "Kaesong-si",
-    "9": "Kangwon-do",
-    "11": "P'yongan-bukto",
-    "12": "P'yongyang-si",
-    "13": "Yanggang-do",
-    "14": "Namp'o-si",
-    "15": "P'yongan-namdo",
-    "17": "Hamgyong-bukto",
-    "18": "Najin Sonbong-si"
-  },
-  "KR": {
-    "1": "Cheju-do",
-    "3": "Cholla-bukto",
-    "5": "Ch'ungch'ong-bukto",
-    "6": "Kangwon-do",
-    "10": "Pusan-jikhalsi",
-    "11": "Seoul-t'ukpyolsi",
-    "12": "Inch'on-jikhalsi",
-    "13": "Kyonggi-do",
-    "14": "Kyongsang-bukto",
-    "15": "Taegu-jikhalsi",
-    "16": "Cholla-namdo",
-    "17": "Ch'ungch'ong-namdo",
-    "18": "Kwangju-jikhalsi",
-    "19": "Taejon-jikhalsi",
-    "20": "Kyongsang-namdo",
-    "21": "Ulsan-gwangyoksi"
-  },
-  "KW": {
-    "1": "Al Ahmadi",
-    "2": "Al Kuwayt",
-    "5": "Al Jahra",
-    "7": "Al Farwaniyah",
-    "8": "Hawalli",
-    "9": "Mubarak al Kabir"
-  },
-  "KY": {
-    "1": "Creek",
-    "2": "Eastern",
-    "3": "Midland",
-    "4": "South Town",
-    "5": "Spot Bay",
-    "6": "Stake Bay",
-    "7": "West End",
-    "8": "Western"
-  },
-  "KZ": {
-    "1": "Almaty",
-    "2": "Almaty City",
-    "3": "Aqmola",
-    "4": "Aqtobe",
-    "5": "Astana",
-    "6": "Atyrau",
-    "7": "West Kazakhstan",
-    "8": "Bayqonyr",
-    "9": "Mangghystau",
-    "10": "South Kazakhstan",
-    "11": "Pavlodar",
-    "12": "Qaraghandy",
-    "13": "Qostanay",
-    "14": "Qyzylorda",
-    "15": "East Kazakhstan",
-    "16": "North Kazakhstan",
-    "17": "Zhambyl"
-  },
-  "LA": {
-    "1": "Attapu",
-    "2": "Champasak",
-    "3": "Houaphan",
-    "4": "Khammouan",
-    "5": "Louang Namtha",
-    "7": "Oudomxai",
-    "8": "Phongsali",
-    "9": "Saravan",
-    "10": "Savannakhet",
-    "11": "Vientiane",
-    "13": "Xaignabouri",
-    "14": "Xiangkhoang",
-    "17": "Louangphrabang"
-  },
-  "LB": {
-    "1": "Beqaa",
-    "2": "Al Janub",
-    "3": "Liban-Nord",
-    "4": "Beyrouth",
-    "5": "Mont-Liban",
-    "6": "Liban-Sud",
-    "7": "Nabatiye",
-    "8": "Beqaa",
-    "9": "Liban-Nord",
-    "10": "Aakk",
-    "11": "Baalbek-Hermel"
-  },
-  "LC": {
-    "1": "Anse-la-Raye",
-    "2": "Dauphin",
-    "3": "Castries",
-    "4": "Choiseul",
-    "5": "Dennery",
-    "6": "Gros-Islet",
-    "7": "Laborie",
-    "8": "Micoud",
-    "9": "Soufriere",
-    "10": "Vieux-Fort",
-    "11": "Praslin"
-  },
-  "LI": {
-    "1": "Balzers",
-    "2": "Eschen",
-    "3": "Gamprin",
-    "4": "Mauren",
-    "5": "Planken",
-    "6": "Ruggell",
-    "7": "Schaan",
-    "8": "Schellenberg",
-    "9": "Triesen",
-    "10": "Triesenberg",
-    "11": "Vaduz",
-    "21": "Gbarpolu",
-    "22": "River Gee"
-  },
-  "LK": {
-    "29": "Central",
-    "30": "North Central",
-    "32": "North Western",
-    "33": "Sabaragamuwa",
-    "34": "Southern",
-    "35": "Uva",
-    "36": "Western",
-    "37": "Eastern",
-    "38": "Northern"
-  },
-  "LR": {
-    "1": "Bong",
-    "4": "Grand Cape Mount",
-    "5": "Lofa",
-    "6": "Maryland",
-    "7": "Monrovia",
-    "9": "Nimba",
-    "10": "Sino",
-    "11": "Grand Bassa",
-    "12": "Grand Cape Mount",
-    "13": "Maryland",
-    "14": "Montserrado",
-    "17": "Margibi",
-    "18": "River Cess",
-    "19": "Grand Gedeh",
-    "20": "Lofa",
-    "21": "Gbarpolu",
-    "22": "River Gee"
-  },
-  "LS": {
-    "10": "Berea",
-    "11": "Butha-Buthe",
-    "12": "Leribe",
-    "13": "Mafeteng",
-    "14": "Maseru",
-    "15": "Mohales Hoek",
-    "16": "Mokhotlong",
-    "17": "Qachas Nek",
-    "18": "Quthing",
-    "19": "Thaba-Tseka"
-  },
-  "LT": {
-    "56": "Alytaus Apskritis",
-    "57": "Kauno Apskritis",
-    "58": "Klaipedos Apskritis",
-    "59": "Marijampoles Apskritis",
-    "60": "Panevezio Apskritis",
-    "61": "Siauliu Apskritis",
-    "62": "Taurages Apskritis",
-    "63": "Telsiu Apskritis",
-    "64": "Utenos Apskritis",
-    "65": "Vilniaus Apskritis"
-  },
-  "LU": {
-    "1": "Diekirch",
-    "2": "Grevenmacher",
-    "3": "Luxembourg"
-  },
-  "LV": {
-    "1": "Aizkraukles",
-    "2": "Aluksnes",
-    "3": "Balvu",
-    "4": "Bauskas",
-    "5": "Cesu",
-    "6": "Daugavpils",
-    "7": "Daugavpils",
-    "8": "Dobeles",
-    "9": "Gulbenes",
-    "10": "Jekabpils",
-    "11": "Jelgava",
-    "12": "Jelgavas",
-    "13": "Jurmala",
-    "14": "Kraslavas",
-    "15": "Kuldigas",
-    "16": "Liepaja",
-    "17": "Liepajas",
-    "18": "Limbazu",
-    "19": "Ludzas",
-    "20": "Madonas",
-    "21": "Ogres",
-    "22": "Preilu",
-    "23": "Rezekne",
-    "24": "Rezeknes",
-    "25": "Riga",
-    "26": "Rigas",
-    "27": "Saldus",
-    "28": "Talsu",
-    "29": "Tukuma",
-    "30": "Valkas",
-    "31": "Valmieras",
-    "32": "Ventspils",
-    "33": "Ventspils"
-  },
-  "LY": {
-    "3": "Al Aziziyah",
-    "5": "Al Jufrah",
-    "8": "Al Kufrah",
-    "13": "Ash Shati'",
-    "30": "Murzuq",
-    "34": "Sabha",
-    "41": "Tarhunah",
-    "42": "Tubruq",
-    "45": "Zlitan",
-    "47": "Ajdabiya",
-    "48": "Al Fatih",
-    "49": "Al Jabal al Akhdar",
-    "50": "Al Khums",
-    "51": "An Nuqat al Khams",
-    "52": "Awbari",
-    "53": "Az Zawiyah",
-    "54": "Banghazi",
-    "55": "Darnah",
-    "56": "Ghadamis",
-    "57": "Gharyan",
-    "58": "Misratah",
-    "59": "Sawfajjin",
-    "60": "Surt",
-    "61": "Tarabulus",
-    "62": "Yafran"
-  },
-  "MA": {
-    "45": "Grand Casablanca",
-    "46": "Fes-Boulemane",
-    "47": "Marrakech-Tensift-Al Haouz",
-    "48": "Meknes-Tafilalet",
-    "49": "Rabat-Sale-Zemmour-Zaer",
-    "50": "Chaouia-Ouardigha",
-    "51": "Doukkala-Abda",
-    "52": "Gharb-Chrarda-Beni Hssen",
-    "53": "Guelmim-Es Smara",
-    "54": "Oriental",
-    "55": "Souss-Massa-Dr",
-    "56": "Tadla-Azilal",
-    "57": "Tanger-Tetouan",
-    "58": "Taza-Al Hoceima-Taounate",
-    "59": "La"
-  },
-  "MC": {
-    "1": "La Condamine",
-    "2": "Monaco",
-    "3": "Monte-Carlo"
-  },
-  "MD": {
-    "51": "Gagauzia",
-    "57": "Chisinau",
-    "58": "Stinga Nistrului",
-    "59": "Anenii Noi",
-    "60": "Balti",
-    "61": "Basarabeasca",
-    "62": "Bender",
-    "63": "Briceni",
-    "64": "Cahul",
-    "65": "Cantemir",
-    "66": "Calarasi",
-    "67": "Causeni",
-    "68": "Cimislia",
-    "69": "Criuleni",
-    "70": "Donduseni",
-    "71": "Drochia",
-    "72": "Dubasari",
-    "73": "Edinet",
-    "74": "Falesti",
-    "75": "Floresti",
-    "76": "Glodeni",
-    "77": "Hincesti",
-    "78": "Ialoveni",
-    "79": "Leova",
-    "80": "Nisporeni",
-    "81": "Ocnita",
-    "82": "Orhei",
-    "83": "Rezina",
-    "84": "Riscani",
-    "85": "Singerei",
-    "86": "Soldanesti",
-    "87": "Soroca",
-    "88": "Stefan-Voda",
-    "89": "Straseni",
-    "90": "Taraclia",
-    "91": "Telenesti",
-    "92": "Ungheni"
-  },
-  "MG": {
-    "1": "Antsiranana",
-    "2": "Fianarantsoa",
-    "3": "Mahajanga",
-    "4": "Toamasina",
-    "5": "Antananarivo",
-    "6": "Toliara"
-  },
-  "MK": {
-    "1": "Aracinovo",
-    "2": "Bac",
-    "3": "Belcista",
-    "4": "Berovo",
-    "5": "Bistrica",
-    "6": "Bitola",
-    "7": "Blatec",
-    "8": "Bogdanci",
-    "9": "Bogomila",
-    "10": "Bogovinje",
-    "11": "Bosilovo",
-    "12": "Brvenica",
-    "13": "Cair",
-    "14": "Capari",
-    "15": "Caska",
-    "16": "Cegrane",
-    "17": "Centar",
-    "18": "Centar Zupa",
-    "19": "Cesinovo",
-    "20": "Cucer-Sandevo",
-    "21": "Debar",
-    "22": "Delcevo",
-    "23": "Delogozdi",
-    "24": "Demir Hisar",
-    "25": "Demir Kapija",
-    "26": "Dobrusevo",
-    "27": "Dolna Banjica",
-    "28": "Dolneni",
-    "29": "Dorce Petrov",
-    "30": "Drugovo",
-    "31": "Dzepciste",
-    "32": "Gazi Baba",
-    "33": "Gevgelija",
-    "34": "Gostivar",
-    "35": "Gradsko",
-    "36": "Ilinden",
-    "37": "Izvor",
-    "38": "Jegunovce",
-    "39": "Kamenjane",
-    "40": "Karbinci",
-    "41": "Karpos",
-    "42": "Kavadarci",
-    "43": "Kicevo",
-    "44": "Kisela Voda",
-    "45": "Klecevce",
-    "46": "Kocani",
-    "47": "Konce",
-    "48": "Kondovo",
-    "49": "Konopiste",
-    "50": "Kosel",
-    "51": "Kratovo",
-    "52": "Kriva Palanka",
-    "53": "Krivogastani",
-    "54": "Krusevo",
-    "55": "Kuklis",
-    "56": "Kukurecani",
-    "57": "Kumanovo",
-    "58": "Labunista",
-    "59": "Lipkovo",
-    "60": "Lozovo",
-    "61": "Lukovo",
-    "62": "Makedonska Kamenica",
-    "63": "Makedonski Brod",
-    "64": "Mavrovi Anovi",
-    "65": "Meseista",
-    "66": "Miravci",
-    "67": "Mogila",
-    "68": "Murtino",
-    "69": "Negotino",
-    "70": "Negotino-Polosko",
-    "71": "Novaci",
-    "72": "Novo Selo",
-    "73": "Oblesevo",
-    "74": "Ohrid",
-    "75": "Orasac",
-    "76": "Orizari",
-    "77": "Oslomej",
-    "78": "Pehcevo",
-    "79": "Petrovec",
-    "80": "Plasnica",
-    "81": "Podares",
-    "82": "Prilep",
-    "83": "Probistip",
-    "84": "Radovis",
-    "85": "Rankovce",
-    "86": "Resen",
-    "87": "Rosoman",
-    "88": "Rostusa",
-    "89": "Samokov",
-    "90": "Saraj",
-    "91": "Sipkovica",
-    "92": "Sopiste",
-    "93": "Sopotnica",
-    "94": "Srbinovo",
-    "95": "Staravina",
-    "96": "Star Dojran",
-    "97": "Staro Nagoricane",
-    "98": "Stip",
-    "99": "Struga",
-    "832": "Strumica",
-    "833": "Studenicani",
-    "834": "Suto Orizari",
-    "835": "Sveti Nikole",
-    "836": "Tearce",
-    "837": "Tetovo",
-    "838": "Topolcani",
-    "839": "Valandovo",
-    "840": "Vasilevo",
-    "875": "Veles",
-    "876": "Velesta",
-    "877": "Vevcani",
-    "878": "Vinica",
-    "879": "Vitoliste",
-    "880": "Vranestica",
-    "881": "Vrapciste",
-    "882": "Vratnica",
-    "883": "Vrutok",
-    "918": "Zajas",
-    "919": "Zelenikovo",
-    "920": "Zelino",
-    "921": "Zitose",
-    "922": "Zletovo",
-    "923": "Zrnovci",
-    "925": "Cair",
-    "926": "Caska",
-    "962": "Debar",
-    "963": "Demir Hisar",
-    "964": "Gostivar",
-    "966": "Kavadarci",
-    "967": "Kumanovo",
-    "968": "Makedonski Brod",
-    "1005": "Ohrid",
-    "1006": "Prilep",
-    "1008": "Dojran",
-    "1009": "Struga",
-    "1010": "Strumica",
-    "1011": "Tetovo",
-    "1012": "Valandovo",
-    "1047": "Veles",
-    "1048": "Aerodrom"
-  },
-  "ML": {
-    "1": "Bamako",
-    "3": "Kayes",
-    "4": "Mopti",
-    "5": "Segou",
-    "6": "Sikasso",
-    "7": "Koulikoro",
-    "8": "Tombouctou",
-    "9": "Gao",
-    "10": "Kidal"
-  },
-  "MM": {
-    "1": "Rakhine State",
-    "2": "Chin State",
-    "3": "Irrawaddy",
-    "4": "Kachin State",
-    "5": "Karan State",
-    "6": "Kayah State",
-    "7": "Magwe",
-    "8": "Mandalay",
-    "9": "Pegu",
-    "10": "Sagaing",
-    "11": "Shan State",
-    "12": "Tenasserim",
-    "13": "Mon State",
-    "14": "Rangoon",
-    "17": "Yangon"
-  },
-  "MN": {
-    "1": "Arhangay",
-    "2": "Bayanhongor",
-    "3": "Bayan-Olgiy",
-    "5": "Darhan",
-    "6": "Dornod",
-    "7": "Dornogovi",
-    "8": "Dundgovi",
-    "9": "Dzavhan",
-    "10": "Govi-Altay",
-    "11": "Hentiy",
-    "12": "Hovd",
-    "13": "Hovsgol",
-    "14": "Omnogovi",
-    "15": "Ovorhangay",
-    "16": "Selenge",
-    "17": "Suhbaatar",
-    "18": "Tov",
-    "19": "Uvs",
-    "20": "Ulaanbaatar",
-    "21": "Bulgan",
-    "22": "Erdenet",
-    "23": "Darhan-Uul",
-    "24": "Govisumber",
-    "25": "Orhon"
-  },
-  "MO": {
-    "1": "Ilhas",
-    "2": "Macau"
-  },
-  "MR": {
-    "1": "Hodh Ech Chargui",
-    "2": "Hodh El Gharbi",
-    "3": "Assaba",
-    "4": "Gorgol",
-    "5": "Brakna",
-    "6": "Trarza",
-    "7": "Adrar",
-    "8": "Dakhlet Nouadhibou",
-    "9": "Tagant",
-    "10": "Guidimaka",
-    "11": "Tiris Zemmour",
-    "12": "Inchiri"
-  },
-  "MS": {
-    "1": "Saint Anthony",
-    "2": "Saint Georges",
-    "3": "Saint Peter"
-  },
-  "MU": {
-    "12": "Black River",
-    "13": "Flacq",
-    "14": "Grand Port",
-    "15": "Moka",
-    "16": "Pamplemousses",
-    "17": "Plaines Wilhems",
-    "18": "Port Louis",
-    "19": "Riviere du Rempart",
-    "20": "Savanne",
-    "21": "Agalega Islands",
-    "22": "Cargados Carajos",
-    "23": "Rodrigues"
-  },
-  "MV": {
-    "1": "Seenu",
-    "5": "Laamu",
-    "30": "Alifu",
-    "31": "Baa",
-    "32": "Dhaalu",
-    "33": "Faafu ",
-    "34": "Gaafu Alifu",
-    "35": "Gaafu Dhaalu",
-    "36": "Haa Alifu",
-    "37": "Haa Dhaalu",
-    "38": "Kaafu",
-    "39": "Lhaviyani",
-    "40": "Maale",
-    "41": "Meemu",
-    "42": "Gnaviyani",
-    "43": "Noonu",
-    "44": "Raa",
-    "45": "Shaviyani",
-    "46": "Thaa",
-    "47": "Vaavu"
-  },
-  "MW": {
-    "2": "Chikwawa",
-    "3": "Chiradzulu",
-    "4": "Chitipa",
-    "5": "Thyolo",
-    "6": "Dedza",
-    "7": "Dowa",
-    "8": "Karonga",
-    "9": "Kasungu",
-    "11": "Lilongwe",
-    "12": "Mangochi",
-    "13": "Mchinji",
-    "15": "Mzimba",
-    "16": "Ntcheu",
-    "17": "Nkhata Bay",
-    "18": "Nkhotakota",
-    "19": "Nsanje",
-    "20": "Ntchisi",
-    "21": "Rumphi",
-    "22": "Salima",
-    "23": "Zomba",
-    "24": "Blantyre",
-    "25": "Mwanza",
-    "26": "Balaka",
-    "27": "Likoma",
-    "28": "Machinga",
-    "29": "Mulanje",
-    "30": "Phalombe"
-  },
-  "MX": {
-    "1": "Aguascalientes",
-    "2": "Baja California",
-    "3": "Baja California Sur",
-    "4": "Campeche",
-    "5": "Chiapas",
-    "6": "Chihuahua",
-    "7": "Coahuila de Zaragoza",
-    "8": "Colima",
-    "9": "Distrito Federal",
-    "10": "Durango",
-    "11": "Guanajuato",
-    "12": "Guerrero",
-    "13": "Hidalgo",
-    "14": "Jalisco",
-    "15": "Mexico",
-    "16": "Michoacan de Ocampo",
-    "17": "Morelos",
-    "18": "Nayarit",
-    "19": "Nuevo Leon",
-    "20": "Oaxaca",
-    "21": "Puebla",
-    "22": "Queretaro de Arteaga",
-    "23": "Quintana Roo",
-    "24": "San Luis Potosi",
-    "25": "Sinaloa",
-    "26": "Sonora",
-    "27": "Tabasco",
-    "28": "Tamaulipas",
-    "29": "Tlaxcala",
-    "30": "Veracruz-Llave",
-    "31": "Yucatan",
-    "32": "Zacatecas"
-  },
-  "MY": {
-    "1": "Johor",
-    "2": "Kedah",
-    "3": "Kelantan",
-    "4": "Melaka",
-    "5": "Negeri Sembilan",
-    "6": "Pahang",
-    "7": "Perak",
-    "8": "Perlis",
-    "9": "Pulau Pinang",
-    "11": "Sarawak",
-    "12": "Selangor",
-    "13": "Terengganu",
-    "14": "Kuala Lumpur",
-    "15": "Labuan",
-    "16": "Sabah",
-    "17": "Putrajaya"
-  },
-  "MZ": {
-    "1": "Cabo Delgado",
-    "2": "Gaza",
-    "3": "Inhambane",
-    "4": "Maputo",
-    "5": "Sofala",
-    "6": "Nampula",
-    "7": "Niassa",
-    "8": "Tete",
-    "9": "Zambezia",
-    "10": "Manica",
-    "11": "Maputo"
-  },
-  "NA": {
-    "1": "Bethanien",
-    "2": "Caprivi Oos",
-    "3": "Boesmanland",
-    "4": "Gobabis",
-    "5": "Grootfontein",
-    "6": "Kaokoland",
-    "7": "Karibib",
-    "8": "Keetmanshoop",
-    "9": "Luderitz",
-    "10": "Maltahohe",
-    "11": "Okahandja",
-    "12": "Omaruru",
-    "13": "Otjiwarongo",
-    "14": "Outjo",
-    "15": "Owambo",
-    "16": "Rehoboth",
-    "17": "Swakopmund",
-    "18": "Tsumeb",
-    "20": "Karasburg",
-    "21": "Windhoek",
-    "22": "Damaraland",
-    "23": "Hereroland Oos",
-    "24": "Hereroland Wes",
-    "25": "Kavango",
-    "26": "Mariental",
-    "27": "Namaland",
-    "28": "Caprivi",
-    "29": "Erongo",
-    "30": "Hardap",
-    "31": "Karas",
-    "32": "Kunene",
-    "33": "Ohangwena",
-    "34": "Okavango",
-    "35": "Omaheke",
-    "36": "Omusati",
-    "37": "Oshana",
-    "38": "Oshikoto",
-    "39": "Otjozondjupa"
-  },
-  "NE": {
-    "1": "Agadez",
-    "2": "Diffa",
-    "3": "Dosso",
-    "4": "Maradi",
-    "5": "Niamey",
-    "6": "Tahoua",
-    "7": "Zinder",
-    "8": "Niamey"
-  },
-  "NG": {
-    "5": "Lagos",
-    "11": "Federal Capital Territory",
-    "16": "Ogun",
-    "21": "Akwa Ibom",
-    "22": "Cross River",
-    "23": "Kaduna",
-    "24": "Katsina",
-    "25": "Anambra",
-    "26": "Benue",
-    "27": "Borno",
-    "28": "Imo",
-    "29": "Kano",
-    "30": "Kwara",
-    "31": "Niger",
-    "32": "Oyo",
-    "35": "Adamawa",
-    "36": "Delta",
-    "37": "Edo",
-    "39": "Jigawa",
-    "40": "Kebbi",
-    "41": "Kogi",
-    "42": "Osun",
-    "43": "Taraba",
-    "44": "Yobe",
-    "45": "Abia",
-    "46": "Bauchi",
-    "47": "Enugu",
-    "48": "Ondo",
-    "49": "Plateau",
-    "50": "Rivers",
-    "51": "Sokoto",
-    "52": "Bayelsa",
-    "53": "Ebonyi",
-    "54": "Ekiti",
-    "55": "Gombe",
-    "56": "Nassarawa",
-    "57": "Zamfara"
-  },
-  "NI": {
-    "1": "Boaco",
-    "2": "Carazo",
-    "3": "Chinandega",
-    "4": "Chontales",
-    "5": "Esteli",
-    "6": "Granada",
-    "7": "Jinotega",
-    "8": "Leon",
-    "9": "Madriz",
-    "10": "Managua",
-    "11": "Masaya",
-    "12": "Matagalpa",
-    "13": "Nueva Segovia",
-    "14": "Rio San Juan",
-    "15": "Rivas",
-    "16": "Zelaya",
-    "17": "Autonoma Atlantico Norte",
-    "18": "Region Autonoma Atlantico Sur"
-  },
-  "NL": {
-    "1": "Drenthe",
-    "2": "Friesland",
-    "3": "Gelderland",
-    "4": "Groningen",
-    "5": "Limburg",
-    "6": "Noord-Brabant",
-    "7": "Noord-Holland",
-    "9": "Utrecht",
-    "10": "Zeeland",
-    "11": "Zuid-Holland",
-    "15": "Overijssel",
-    "16": "Flevoland"
-  },
-  "NO": {
-    "1": "Akershus",
-    "2": "Aust-Agder",
-    "4": "Buskerud",
-    "5": "Finnmark",
-    "6": "Hedmark",
-    "7": "Hordaland",
-    "8": "More og Romsdal",
-    "9": "Nordland",
-    "10": "Nord-Trondelag",
-    "11": "Oppland",
-    "12": "Oslo",
-    "13": "Ostfold",
-    "14": "Rogaland",
-    "15": "Sogn og Fjordane",
-    "16": "Sor-Trondelag",
-    "17": "Telemark",
-    "18": "Troms",
-    "19": "Vest-Agder",
-    "20": "Vestfold"
-  },
-  "NP": {
-    "1": "Bagmati",
-    "2": "Bheri",
-    "3": "Dhawalagiri",
-    "4": "Gandaki",
-    "5": "Janakpur",
-    "6": "Karnali",
-    "7": "Kosi",
-    "8": "Lumbini",
-    "9": "Mahakali",
-    "10": "Mechi",
-    "11": "Narayani",
-    "12": "Rapti",
-    "13": "Sagarmatha",
-    "14": "Seti"
-  },
-  "NR": {
-    "1": "Aiwo",
-    "2": "Anabar",
-    "3": "Anetan",
-    "4": "Anibare",
-    "5": "Baiti",
-    "6": "Boe",
-    "7": "Buada",
-    "8": "Denigomodu",
-    "9": "Ewa",
-    "10": "Ijuw",
-    "11": "Meneng",
-    "12": "Nibok",
-    "13": "Uaboe",
-    "14": "Yaren"
-  },
-  "NZ": {
-    "10": "Chatham Islands",
-    "1010": "Auckland",
-    "1011": "Bay of Plenty",
-    "1012": "Canterbury",
-    "1047": "Gisborne",
-    "1048": "Hawke's Bay",
-    "1049": "Manawatu-Wanganui",
-    "1050": "Marlborough",
-    "1051": "Nelson",
-    "1052": "Northland",
-    "1053": "Otago",
-    "1054": "Southland",
-    "1055": "Taranaki",
-    "1090": "Waikato",
-    "1091": "Wellington",
-    "1092": "West Coast"
-  },
-  "OM": {
-    "1": "Ad Dakhiliyah",
-    "2": "Al Batinah",
-    "3": "Al Wusta",
-    "4": "Ash Sharqiyah",
-    "5": "Az Zahirah",
-    "6": "Masqat",
-    "7": "Musandam",
-    "8": "Zufar"
-  },
-  "PA": {
-    "1": "Bocas del Toro",
-    "2": "Chiriqui",
-    "3": "Cocle",
-    "4": "Colon",
-    "5": "Darien",
-    "6": "Herrera",
-    "7": "Los Santos",
-    "8": "Panama",
-    "9": "San Blas",
-    "10": "Veraguas"
-  },
-  "PE": {
-    "1": "Amazonas",
-    "2": "Ancash",
-    "3": "Apurimac",
-    "4": "Arequipa",
-    "5": "Ayacucho",
-    "6": "Cajamarca",
-    "7": "Callao",
-    "8": "Cusco",
-    "9": "Huancavelica",
-    "10": "Huanuco",
-    "11": "Ica",
-    "12": "Junin",
-    "13": "La Libertad",
-    "14": "Lambayeque",
-    "15": "Lima",
-    "16": "Loreto",
-    "17": "Madre de Dios",
-    "18": "Moquegua",
-    "19": "Pasco",
-    "20": "Piura",
-    "21": "Puno",
-    "22": "San Martin",
-    "23": "Tacna",
-    "24": "Tumbes",
-    "25": "Ucayali"
-  },
-  "PG": {
-    "1": "Central",
-    "2": "Gulf",
-    "3": "Milne Bay",
-    "4": "Northern",
-    "5": "Southern Highlands",
-    "6": "Western",
-    "7": "North Solomons",
-    "8": "Chimbu",
-    "9": "Eastern Highlands",
-    "10": "East New Britain",
-    "11": "East Sepik",
-    "12": "Madang",
-    "13": "Manus",
-    "14": "Morobe",
-    "15": "New Ireland",
-    "16": "Western Highlands",
-    "17": "West New Britain",
-    "18": "Sandaun",
-    "19": "Enga",
-    "20": "National Capital"
-  },
-  "PH": {
-    "1": "Abra",
-    "2": "Agusan del Norte",
-    "3": "Agusan del Sur",
-    "4": "Aklan",
-    "5": "Albay",
-    "6": "Antique",
-    "7": "Bataan",
-    "8": "Batanes",
-    "9": "Batangas",
-    "10": "Benguet",
-    "11": "Bohol",
-    "12": "Bukidnon",
-    "13": "Bulacan",
-    "14": "Cagayan",
-    "15": "Camarines Norte",
-    "16": "Camarines Sur",
-    "17": "Camiguin",
-    "18": "Capiz",
-    "19": "Catanduanes",
-    "20": "Cavite",
-    "21": "Cebu",
-    "22": "Basilan",
-    "23": "Eastern Samar",
-    "24": "Davao",
-    "25": "Davao del Sur",
-    "26": "Davao Oriental",
-    "27": "Ifugao",
-    "28": "Ilocos Norte",
-    "29": "Ilocos Sur",
-    "30": "Iloilo",
-    "31": "Isabela",
-    "32": "Kalinga-Apayao",
-    "33": "Laguna",
-    "34": "Lanao del Norte",
-    "35": "Lanao del Sur",
-    "36": "La Union",
-    "37": "Leyte",
-    "38": "Marinduque",
-    "39": "Masbate",
-    "40": "Mindoro Occidental",
-    "41": "Mindoro Oriental",
-    "42": "Misamis Occidental",
-    "43": "Misamis Oriental",
-    "44": "Mountain",
-    "45": "Negros Occidental",
-    "46": "Negros Oriental",
-    "47": "Nueva Ecija",
-    "48": "Nueva Vizcaya",
-    "49": "Palawan",
-    "50": "Pampanga",
-    "51": "Pangasinan",
-    "53": "Rizal",
-    "54": "Romblon",
-    "55": "Samar",
-    "56": "Maguindanao",
-    "57": "North Cotabato",
-    "58": "Sorsogon",
-    "59": "Southern Leyte",
-    "60": "Sulu",
-    "61": "Surigao del Norte",
-    "62": "Surigao del Sur",
-    "63": "Tarlac",
-    "64": "Zambales",
-    "65": "Zamboanga del Norte",
-    "66": "Zamboanga del Sur",
-    "67": "Northern Samar",
-    "68": "Quirino",
-    "69": "Siquijor",
-    "70": "South Cotabato",
-    "71": "Sultan Kudarat",
-    "72": "Tawitawi",
-    "832": "Angeles",
-    "833": "Bacolod",
-    "834": "Bago",
-    "835": "Baguio",
-    "836": "Bais",
-    "837": "Basilan City",
-    "838": "Batangas City",
-    "839": "Butuan",
-    "840": "Cabanatuan",
-    "875": "Cadiz",
-    "876": "Cagayan de Oro",
-    "877": "Calbayog",
-    "878": "Caloocan",
-    "879": "Canlaon",
-    "880": "Cavite City",
-    "881": "Cebu City",
-    "882": "Cotabato",
-    "883": "Dagupan",
-    "918": "Danao",
-    "919": "Dapitan",
-    "920": "Davao City",
-    "921": "Dipolog",
-    "922": "Dumaguete",
-    "923": "General Santos",
-    "924": "Gingoog",
-    "925": "Iligan",
-    "926": "Iloilo City",
-    "961": "Iriga",
-    "962": "La Carlota",
-    "963": "Laoag",
-    "964": "Lapu-Lapu",
-    "965": "Legaspi",
-    "966": "Lipa",
-    "967": "Lucena",
-    "968": "Mandaue",
-    "969": "Manila",
-    "1004": "Marawi",
-    "1005": "Naga",
-    "1006": "Olongapo",
-    "1007": "Ormoc",
-    "1008": "Oroquieta",
-    "1009": "Ozamis",
-    "1010": "Pagadian",
-    "1011": "Palayan",
-    "1012": "Pasay",
-    "1047": "Puerto Princesa",
-    "1048": "Quezon City",
-    "1049": "Roxas",
-    "1050": "San Carlos",
-    "1051": "San Carlos",
-    "1052": "San Jose",
-    "1053": "San Pablo",
-    "1054": "Silay",
-    "1055": "Surigao",
-    "1090": "Tacloban",
-    "1091": "Tagaytay",
-    "1092": "Tagbilaran",
-    "1093": "Tangub",
-    "1094": "Toledo",
-    "1095": "Trece Martires",
-    "1096": "Zamboanga",
-    "1097": "Aurora",
-    "1134": "Quezon",
-    "1135": "Negros Occidental",
-    "1141": "Biliran",
-    "1181": "Compostela Valley",
-    "1182": "Davao del Norte",
-    "1221": "Guimaras",
-    "1222": "Himamaylan",
-    "1225": "Kalinga",
-    "1262": "Las Pinas",
-    "1266": "Malabon",
-    "1267": "Malaybalay",
-    "1308": "Muntinlupa",
-    "1309": "Navotas",
-    "1311": "Paranaque",
-    "1313": "Passi",
-    "1477": "Zambales",
-    "1352": "San Jose del Monte",
-    "1353": "San Juan",
-    "1355": "Santiago",
-    "1356": "Sarangani",
-    "1391": "Sipalay",
-    "1393": "Surigao del Norte",
-    "1478": "Zamboanga"
-  },
-  "PK": {
-    "1": "Federally Administered Tribal Areas",
-    "2": "Balochistan",
-    "3": "North-West Frontier",
-    "4": "Punjab",
-    "5": "Sindh",
-    "6": "Azad Kashmir",
-    "7": "Northern Areas",
-    "8": "Islamabad"
-  },
-  "PL": {
-    "72": "Dolnoslaskie",
-    "73": "Kujawsko-Pomorskie",
-    "74": "Lodzkie",
-    "75": "Lubelskie",
-    "76": "Lubuskie",
-    "77": "Malopolskie",
-    "78": "Mazowieckie",
-    "79": "Opolskie",
-    "80": "Podkarpackie",
-    "81": "Podlaskie",
-    "82": "Pomorskie",
-    "83": "Slaskie",
-    "84": "Swietokrzyskie",
-    "85": "Warminsko-Mazurskie",
-    "86": "Wielkopolskie",
-    "87": "Zachodniopomorskie"
-  },
-  "PS": {
-    "1131": "Gaza",
-    "1798": "West Bank"
-  },
-  "PT": {
-    "2": "Aveiro",
-    "3": "Beja",
-    "4": "Braga",
-    "5": "Braganca",
-    "6": "Castelo Branco",
-    "7": "Coimbra",
-    "8": "Evora",
-    "9": "Faro",
-    "10": "Madeira",
-    "11": "Guarda",
-    "13": "Leiria",
-    "14": "Lisboa",
-    "16": "Portalegre",
-    "17": "Porto",
-    "18": "Santarem",
-    "19": "Setubal",
-    "20": "Viana do Castelo",
-    "21": "Vila Real",
-    "22": "Viseu",
-    "23": "Azores"
-  },
-  "PY": {
-    "1": "Alto Parana",
-    "2": "Amambay",
-    "4": "Caaguazu",
-    "5": "Caazapa",
-    "6": "Central",
-    "7": "Concepcion",
-    "8": "Cordillera",
-    "10": "Guaira",
-    "11": "Itapua",
-    "12": "Misiones",
-    "13": "Neembucu",
-    "15": "Paraguari",
-    "16": "Presidente Hayes",
-    "17": "San Pedro",
-    "19": "Canindeyu",
-    "22": "Asuncion",
-    "23": "Alto Paraguay",
-    "24": "Boqueron"
-  },
-  "QA": {
-    "1": "Ad Dawhah",
-    "2": "Al Ghuwariyah",
-    "3": "Al Jumaliyah",
-    "4": "Al Khawr",
-    "5": "Al Wakrah Municipality",
-    "6": "Ar Rayyan",
-    "8": "Madinat ach Shamal",
-    "9": "Umm Salal",
-    "10": "Al Wakrah",
-    "11": "Jariyan al Batnah",
-    "12": "Umm Sa'id"
-  },
-  "RO": {
-    "1": "Alba",
-    "2": "Arad",
-    "3": "Arges",
-    "4": "Bacau",
-    "5": "Bihor",
-    "6": "Bistrita-Nasaud",
-    "7": "Botosani",
-    "8": "Braila",
-    "9": "Brasov",
-    "10": "Bucuresti",
-    "11": "Buzau",
-    "12": "Caras-Severin",
-    "13": "Cluj",
-    "14": "Constanta",
-    "15": "Covasna",
-    "16": "Dambovita",
-    "17": "Dolj",
-    "18": "Galati",
-    "19": "Gorj",
-    "20": "Harghita",
-    "21": "Hunedoara",
-    "22": "Ialomita",
-    "23": "Iasi",
-    "25": "Maramures",
-    "26": "Mehedinti",
-    "27": "Mures",
-    "28": "Neamt",
-    "29": "Olt",
-    "30": "Prahova",
-    "31": "Salaj",
-    "32": "Satu Mare",
-    "33": "Sibiu",
-    "34": "Suceava",
-    "35": "Teleorman",
-    "36": "Timis",
-    "37": "Tulcea",
-    "38": "Vaslui",
-    "39": "Valcea",
-    "40": "Vrancea",
-    "41": "Calarasi",
-    "42": "Giurgiu",
-    "43": "Ilfov"
-  },
-  "RS": {
-    "1": "Kosovo",
-    "2": "Vojvodina"
-  },
-  "RU": {
-    "1": "Adygeya",
-    "2": "Aginsky Buryatsky AO",
-    "3": "Gorno-Altay",
-    "4": "Altaisky krai",
-    "5": "Amur",
-    "6": "Arkhangel'sk",
-    "7": "Astrakhan'",
-    "8": "Bashkortostan",
-    "9": "Belgorod",
-    "10": "Bryansk",
-    "11": "Buryat",
-    "12": "Chechnya",
-    "13": "Chelyabinsk",
-    "14": "Chita",
-    "15": "Chukot",
-    "16": "Chuvashia",
-    "17": "Dagestan",
-    "18": "Evenk",
-    "19": "Ingush",
-    "20": "Irkutsk",
-    "21": "Ivanovo",
-    "22": "Kabardin-Balkar",
-    "23": "Kaliningrad",
-    "24": "Kalmyk",
-    "25": "Kaluga",
-    "26": "Kamchatka",
-    "27": "Karachay-Cherkess",
-    "28": "Karelia",
-    "29": "Kemerovo",
-    "30": "Khabarovsk",
-    "31": "Khakass",
-    "32": "Khanty-Mansiy",
-    "33": "Kirov",
-    "34": "Komi",
-    "36": "Koryak",
-    "37": "Kostroma",
-    "38": "Krasnodar",
-    "39": "Krasnoyarsk",
-    "40": "Kurgan",
-    "41": "Kursk",
-    "42": "Leningrad",
-    "43": "Lipetsk",
-    "44": "Magadan",
-    "45": "Mariy-El",
-    "46": "Mordovia",
-    "47": "Moskva",
-    "48": "Moscow City",
-    "49": "Murmansk",
-    "50": "Nenets",
-    "51": "Nizhegorod",
-    "52": "Novgorod",
-    "53": "Novosibirsk",
-    "54": "Omsk",
-    "55": "Orenburg",
-    "56": "Orel",
-    "57": "Penza",
-    "58": "Perm'",
-    "59": "Primor'ye",
-    "60": "Pskov",
-    "61": "Rostov",
-    "62": "Ryazan'",
-    "63": "Sakha",
-    "64": "Sakhalin",
-    "65": "Samara",
-    "66": "Saint Petersburg City",
-    "67": "Saratov",
-    "68": "North Ossetia",
-    "69": "Smolensk",
-    "70": "Stavropol'",
-    "71": "Sverdlovsk",
-    "72": "Tambovskaya oblast",
-    "73": "Tatarstan",
-    "74": "Taymyr",
-    "75": "Tomsk",
-    "76": "Tula",
-    "77": "Tver'",
-    "78": "Tyumen'",
-    "79": "Tuva",
-    "80": "Udmurt",
-    "81": "Ul'yanovsk",
-    "83": "Vladimir",
-    "84": "Volgograd",
-    "85": "Vologda",
-    "86": "Voronezh",
-    "87": "Yamal-Nenets",
-    "88": "Yaroslavl'",
-    "89": "Yevrey",
-    "90": "Permskiy Kray",
-    "91": "Krasnoyarskiy Kray",
-    "92": "Kamchatskiy Kray",
-    "93": "Zabaykal'skiy Kray"
-  },
-  "RW": {
-    "1": "Butare",
-    "6": "Gitarama",
-    "7": "Kibungo",
-    "9": "Kigali",
-    "11": "Est",
-    "12": "Kigali",
-    "13": "Nord",
-    "14": "Ouest",
-    "15": "Sud"
-  },
-  "SA": {
-    "2": "Al Bahah",
-    "5": "Al Madinah",
-    "6": "Ash Sharqiyah",
-    "8": "Al Qasim",
-    "10": "Ar Riyad",
-    "11": "Asir Province",
-    "13": "Ha'il",
-    "14": "Makkah",
-    "15": "Al Hudud ash Shamaliyah",
-    "16": "Najran",
-    "17": "Jizan",
-    "19": "Tabuk",
-    "20": "Al Jawf"
-  },
-  "SB": {
-    "3": "Malaita",
-    "6": "Guadalcanal",
-    "7": "Isabel",
-    "8": "Makira",
-    "9": "Temotu",
-    "10": "Central",
-    "11": "Western",
-    "12": "Choiseul",
-    "13": "Rennell and Bellona"
-  },
-  "SC": {
-    "1": "Anse aux Pins",
-    "2": "Anse Boileau",
-    "3": "Anse Etoile",
-    "4": "Anse Louis",
-    "5": "Anse Royale",
-    "6": "Baie Lazare",
-    "7": "Baie Sainte Anne",
-    "8": "Beau Vallon",
-    "9": "Bel Air",
-    "10": "Bel Ombre",
-    "11": "Cascade",
-    "12": "Glacis",
-    "13": "Grand' Anse",
-    "14": "Grand' Anse",
-    "15": "La Digue",
-    "16": "La Riviere Anglaise",
-    "17": "Mont Buxton",
-    "18": "Mont Fleuri",
-    "19": "Plaisance",
-    "20": "Pointe La Rue",
-    "21": "Port Glaud",
-    "22": "Saint Louis",
-    "23": "Takamaka"
-  },
-  "SD": {
-    "27": "Al Wusta",
-    "28": "Al Istiwa'iyah",
-    "29": "Al Khartum",
-    "30": "Ash Shamaliyah",
-    "31": "Ash Sharqiyah",
-    "32": "Bahr al Ghazal",
-    "33": "Darfur",
-    "34": "Kurdufan",
-    "35": "Upper Nile",
-    "40": "Al Wahadah State",
-    "44": "Central Equatoria State",
-    "49": "Southern Darfur",
-    "50": "Southern Kordofan",
-    "52": "Kassala",
-    "53": "River Nile",
-    "55": "Northern Darfur"
-  },
-  "SE": {
-    "2": "Blekinge Lan",
-    "3": "Gavleborgs Lan",
-    "5": "Gotlands Lan",
-    "6": "Hallands Lan",
-    "7": "Jamtlands Lan",
-    "8": "Jonkopings Lan",
-    "9": "Kalmar Lan",
-    "10": "Dalarnas Lan",
-    "12": "Kronobergs Lan",
-    "14": "Norrbottens Lan",
-    "15": "Orebro Lan",
-    "16": "Ostergotlands Lan",
-    "18": "Sodermanlands Lan",
-    "21": "Uppsala Lan",
-    "22": "Varmlands Lan",
-    "23": "Vasterbottens Lan",
-    "24": "Vasternorrlands Lan",
-    "25": "Vastmanlands Lan",
-    "26": "Stockholms Lan",
-    "27": "Skane Lan",
-    "28": "Vastra Gotaland"
-  },
-  "SH": {
-    "1": "Ascension",
-    "2": "Saint Helena",
-    "3": "Tristan da Cunha"
-  },
-  "SI": {
-    "1": "Ajdovscina Commune",
-    "2": "Beltinci Commune",
-    "3": "Bled Commune",
-    "4": "Bohinj Commune",
-    "5": "Borovnica Commune",
-    "6": "Bovec Commune",
-    "7": "Brda Commune",
-    "8": "Brezice Commune",
-    "9": "Brezovica Commune",
-    "11": "Celje Commune",
-    "12": "Cerklje na Gorenjskem Commune",
-    "13": "Cerknica Commune",
-    "14": "Cerkno Commune",
-    "15": "Crensovci Commune",
-    "16": "Crna na Koroskem Commune",
-    "17": "Crnomelj Commune",
-    "19": "Divaca Commune",
-    "20": "Dobrepolje Commune",
-    "22": "Dol pri Ljubljani Commune",
-    "24": "Dornava Commune",
-    "25": "Dravograd Commune",
-    "26": "Duplek Commune",
-    "27": "Gorenja vas-Poljane Commune",
-    "28": "Gorisnica Commune",
-    "29": "Gornja Radgona Commune",
-    "30": "Gornji Grad Commune",
-    "31": "Gornji Petrovci Commune",
-    "32": "Grosuplje Commune",
-    "34": "Hrastnik Commune",
-    "35": "Hrpelje-Kozina Commune",
-    "36": "Idrija Commune",
-    "37": "Ig Commune",
-    "38": "Ilirska Bistrica Commune",
-    "39": "Ivancna Gorica Commune",
-    "40": "Izola-Isola Commune",
-    "42": "Jursinci Commune",
-    "44": "Kanal Commune",
-    "45": "Kidricevo Commune",
-    "46": "Kobarid Commune",
-    "47": "Kobilje Commune",
-    "49": "Komen Commune",
-    "50": "Koper-Capodistria Urban Commune",
-    "51": "Kozje Commune",
-    "52": "Kranj Commune",
-    "53": "Kranjska Gora Commune",
-    "54": "Krsko Commune",
-    "55": "Kungota Commune",
-    "57": "Lasko Commune",
-    "61": "Ljubljana Urban Commune",
-    "62": "Ljubno Commune",
-    "64": "Logatec Commune",
-    "66": "Loski Potok Commune",
-    "68": "Lukovica Commune",
-    "71": "Medvode Commune",
-    "72": "Menges Commune",
-    "73": "Metlika Commune",
-    "74": "Mezica Commune",
-    "76": "Mislinja Commune",
-    "77": "Moravce Commune",
-    "78": "Moravske Toplice Commune",
-    "79": "Mozirje Commune",
-    "80": "Murska Sobota Urban Commune",
-    "81": "Muta Commune",
-    "82": "Naklo Commune",
-    "83": "Nazarje Commune",
-    "84": "Nova Gorica Urban Commune",
-    "86": "Odranci Commune",
-    "87": "Ormoz Commune",
-    "88": "Osilnica Commune",
-    "89": "Pesnica Commune",
-    "91": "Pivka Commune",
-    "92": "Podcetrtek Commune",
-    "94": "Postojna Commune",
-    "97": "Puconci Commune",
-    "98": "Race-Fram Commune",
-    "99": "Radece Commune",
-    "832": "Radenci Commune",
-    "833": "Radlje ob Dravi Commune",
-    "834": "Radovljica Commune",
-    "837": "Rogasovci Commune",
-    "838": "Rogaska Slatina Commune",
-    "839": "Rogatec Commune",
-    "875": "Semic Commune",
-    "876": "Sencur Commune",
-    "877": "Sentilj Commune",
-    "878": "Sentjernej Commune",
-    "880": "Sevnica Commune",
-    "881": "Sezana Commune",
-    "882": "Skocjan Commune",
-    "883": "Skofja Loka Commune",
-    "918": "Skofljica Commune",
-    "919": "Slovenj Gradec Urban Commune",
-    "921": "Slovenske Konjice Commune",
-    "922": "Smarje pri Jelsah Commune",
-    "923": "Smartno ob Paki Commune",
-    "924": "Sostanj Commune",
-    "925": "Starse Commune",
-    "926": "Store Commune",
-    "961": "Sveti Jurij Commune",
-    "962": "Tolmin Commune",
-    "963": "Trbovlje Commune",
-    "964": "Trebnje Commune",
-    "965": "Trzic Commune",
-    "966": "Turnisce Commune",
-    "967": "Velenje Urban Commune",
-    "968": "Velike Lasce Commune",
-    "1004": "Vipava Commune",
-    "1005": "Vitanje Commune",
-    "1006": "Vodice Commune",
-    "1008": "Vrhnika Commune",
-    "1009": "Vuzenica Commune",
-    "1010": "Zagorje ob Savi Commune",
-    "1012": "Zavrc Commune",
-    "1047": "Zelezniki Commune",
-    "1048": "Ziri Commune",
-    "1049": "Zrece Commune",
-    "1050": "Benedikt Commune",
-    "1051": "Bistrica ob Sotli Commune",
-    "1052": "Bloke Commune",
-    "1053": "Braslovce Commune",
-    "1054": "Cankova Commune",
-    "1055": "Cerkvenjak Commune",
-    "1090": "Destrnik Commune",
-    "1091": "Dobje Commune",
-    "1092": "Dobrna Commune",
-    "1093": "Dobrova-Horjul-Polhov Gradec Commune",
-    "1094": "Dobrovnik-Dobronak Commune",
-    "1095": "Dolenjske Toplice Commune",
-    "1096": "Domzale Commune",
-    "1097": "Grad Commune",
-    "1098": "Hajdina Commune",
-    "1133": "Hoce-Slivnica Commune",
-    "1134": "Hodos-Hodos Commune",
-    "1135": "Horjul Commune",
-    "1136": "Jesenice Commune",
-    "1137": "Jezersko Commune",
-    "1138": "Kamnik Commune",
-    "1139": "Kocevje Commune",
-    "1140": "Komenda Commune",
-    "1141": "Kostel Commune",
-    "1176": "Krizevci Commune",
-    "1177": "Kuzma Commune",
-    "1178": "Lenart Commune",
-    "1179": "Lendava-Lendva Commune",
-    "1180": "Litija Commune",
-    "1181": "Ljutomer Commune",
-    "1182": "Loska Dolina Commune",
-    "1183": "Lovrenc na Pohorju Commune",
-    "1184": "Luce Commune",
-    "1219": "Majsperk Commune",
-    "1220": "Maribor Commune",
-    "1221": "Markovci Commune",
-    "1222": "Miklavz na Dravskem polju Commune",
-    "1223": "Miren-Kostanjevica Commune",
-    "1224": "Mirna Pec Commune",
-    "1225": "Novo mesto Urban Commune",
-    "1226": "Oplotnica Commune",
-    "1227": "Piran-Pirano Commune",
-    "1262": "Podlehnik Commune",
-    "1263": "Podvelka Commune",
-    "1264": "Polzela Commune",
-    "1265": "Prebold Commune",
-    "1266": "Preddvor Commune",
-    "1267": "Prevalje Commune",
-    "1268": "Ptuj Urban Commune",
-    "1269": "Ravne na Koroskem Commune",
-    "1270": "Razkrizje Commune",
-    "1305": "Ribnica Commune",
-    "1306": "Ribnica na Pohorju Commune",
-    "1307": "Ruse Commune",
-    "1308": "Salovci Commune",
-    "1309": "Selnica ob Dravi Commune",
-    "1310": "Sempeter-Vrtojba Commune",
-    "1311": "Sentjur pri Celju Commune",
-    "1312": "Slovenska Bistrica Commune",
-    "1313": "Smartno pri Litiji Commune",
-    "1348": "Sodrazica Commune",
-    "1349": "Solcava Commune",
-    "1350": "Sveta Ana Commune",
-    "1351": "Sveti Andraz v Slovenskih goricah Commune",
-    "1352": "Tabor Commune",
-    "1353": "Tisina Commune",
-    "1354": "Trnovska vas Commune",
-    "1355": "Trzin Commune",
-    "1356": "Velika Polana Commune",
-    "1391": "Verzej Commune",
-    "1392": "Videm Commune",
-    "1393": "Vojnik Commune",
-    "1394": "Vransko Commune",
-    "1395": "Zalec Commune",
-    "1396": "Zetale Commune",
-    "1397": "Zirovnica Commune",
-    "1398": "Zuzemberk Commune",
-    "1399": "Apace Commune",
-    "1434": "Cirkulane Commune",
-    "1435": "Gorje",
-    "1436": "Kostanjevica na Krki",
-    "1437": "Log-Dragomer",
-    "1438": "Makole",
-    "1439": "Mirna",
-    "1440": "Mokronog-Trebelno",
-    "1441": "Poljcane",
-    "1442": "Recica ob Savinji",
-    "1477": "Rence-Vogrsko",
-    "1478": "Sentrupert",
-    "1479": "Smarjesk Toplice",
-    "1480": "Sredisce ob Dravi",
-    "1481": "Straza",
-    "1483": "Sveti Jurij v Slovenskih Goricah"
-  },
-  "SK": {
-    "1": "Banska Bystrica",
-    "2": "Bratislava",
-    "3": "Kosice",
-    "4": "Nitra",
-    "5": "Presov",
-    "6": "Trencin",
-    "7": "Trnava",
-    "8": "Zilina"
-  },
-  "SL": {
-    "1": "Eastern",
-    "2": "Northern",
-    "3": "Southern",
-    "4": "Western Area"
-  },
-  "SM": {
-    "1": "Acquaviva",
-    "2": "Chiesanuova",
-    "3": "Domagnano",
-    "4": "Faetano",
-    "5": "Fiorentino",
-    "6": "Borgo Maggiore",
-    "7": "San Marino",
-    "8": "Monte Giardino",
-    "9": "Serravalle"
-  },
-  "SN": {
-    "1": "Dakar",
-    "3": "Diourbel",
-    "5": "Tambacounda",
-    "7": "Thies",
-    "9": "Fatick",
-    "10": "Kaolack",
-    "11": "Kolda",
-    "12": "Ziguinchor",
-    "13": "Louga",
-    "14": "Saint-Louis",
-    "15": "Matam"
-  },
-  "SO": {
-    "1": "Bakool",
-    "2": "Banaadir",
-    "3": "Bari",
-    "4": "Bay",
-    "5": "Galguduud",
-    "6": "Gedo",
-    "7": "Hiiraan",
-    "8": "Jubbada Dhexe",
-    "9": "Jubbada Hoose",
-    "10": "Mudug",
-    "11": "Nugaal",
-    "12": "Sanaag",
-    "13": "Shabeellaha Dhexe",
-    "14": "Shabeellaha Hoose",
-    "16": "Woqooyi Galbeed",
-    "18": "Nugaal",
-    "19": "Togdheer",
-    "20": "Woqooyi Galbeed",
-    "21": "Awdal",
-    "22": "Sool"
-  },
-  "SR": {
-    "10": "Brokopondo",
-    "11": "Commewijne",
-    "12": "Coronie",
-    "13": "Marowijne",
-    "14": "Nickerie",
-    "15": "Para",
-    "16": "Paramaribo",
-    "17": "Saramacca",
-    "18": "Sipaliwini",
-    "19": "Wanica"
-  },
-  "SS": {
-    "1": "Central Equatoria",
-    "2": "Eastern Equatoria",
-    "3": "Jonglei",
-    "4": "Lakes",
-    "5": "Northern Bahr el Ghazal",
-    "6": "Unity",
-    "7": "Upper Nile",
-    "8": "Warrap",
-    "9": "Western Bahr el Ghazal",
-    "10": "Western Equatoria"
-  },
-  "ST": {
-    "1": "Principe",
-    "2": "Sao Tome"
-  },
-  "SV": {
-    "1": "Ahuachapan",
-    "2": "Cabanas",
-    "3": "Chalatenango",
-    "4": "Cuscatlan",
-    "5": "La Libertad",
-    "6": "La Paz",
-    "7": "La Union",
-    "8": "Morazan",
-    "9": "San Miguel",
-    "10": "San Salvador",
-    "11": "Santa Ana",
-    "12": "San Vicente",
-    "13": "Sonsonate",
-    "14": "Usulutan"
-  },
-  "SY": {
-    "1": "Al Hasakah",
-    "2": "Al Ladhiqiyah",
-    "3": "Al Qunaytirah",
-    "4": "Ar Raqqah",
-    "5": "As Suwayda'",
-    "6": "Dar",
-    "7": "Dayr az Zawr",
-    "8": "Rif Dimashq",
-    "9": "Halab",
-    "10": "Hamah",
-    "11": "Hims",
-    "12": "Idlib",
-    "13": "Dimashq",
-    "14": "Tartus"
-  },
-  "SZ": {
-    "1": "Hhohho",
-    "2": "Lubombo",
-    "3": "Manzini",
-    "4": "Shiselweni",
-    "5": "Praslin"
-  },
-  "TD": {
-    "1": "Batha",
-    "2": "Biltine",
-    "3": "Borkou-Ennedi-Tibesti",
-    "4": "Chari-Baguirmi",
-    "5": "Guera",
-    "6": "Kanem",
-    "7": "Lac",
-    "8": "Logone Occidental",
-    "9": "Logone Oriental",
-    "10": "Mayo-Kebbi",
-    "11": "Moyen-Chari",
-    "12": "Ouaddai",
-    "13": "Salamat",
-    "14": "Tandjile"
-  },
-  "TG": {
-    "22": "Centrale",
-    "23": "Kara",
-    "24": "Maritime",
-    "25": "Plateaux",
-    "26": "Savanes"
-  },
-  "TH": {
-    "1": "Mae Hong Son",
-    "2": "Chiang Mai",
-    "3": "Chiang Rai",
-    "4": "Nan",
-    "5": "Lamphun",
-    "6": "Lampang",
-    "7": "Phrae",
-    "8": "Tak",
-    "9": "Sukhothai",
-    "10": "Uttaradit",
-    "11": "Kamphaeng Phet",
-    "12": "Phitsanulok",
-    "13": "Phichit",
-    "14": "Phetchabun",
-    "15": "Uthai Thani",
-    "16": "Nakhon Sawan",
-    "17": "Nong Khai",
-    "18": "Loei",
-    "20": "Sakon Nakhon",
-    "21": "Nakhon Phanom",
-    "22": "Khon Kaen",
-    "23": "Kalasin",
-    "24": "Maha Sarakham",
-    "25": "Roi Et",
-    "26": "Chaiyaphum",
-    "27": "Nakhon Ratchasima",
-    "28": "Buriram",
-    "29": "Surin",
-    "30": "Sisaket",
-    "31": "Narathiwat",
-    "32": "Chai Nat",
-    "33": "Sing Buri",
-    "34": "Lop Buri",
-    "35": "Ang Thong",
-    "36": "Phra Nakhon Si Ayutthaya",
-    "37": "Saraburi",
-    "38": "Nonthaburi",
-    "39": "Pathum Thani",
-    "40": "Krung Thep",
-    "41": "Phayao",
-    "42": "Samut Prakan",
-    "43": "Nakhon Nayok",
-    "44": "Chachoengsao",
-    "45": "Prachin Buri",
-    "46": "Chon Buri",
-    "47": "Rayong",
-    "48": "Chanthaburi",
-    "49": "Trat",
-    "50": "Kanchanaburi",
-    "51": "Suphan Buri",
-    "52": "Ratchaburi",
-    "53": "Nakhon Pathom",
-    "54": "Samut Songkhram",
-    "55": "Samut Sakhon",
-    "56": "Phetchaburi",
-    "57": "Prachuap Khiri Khan",
-    "58": "Chumphon",
-    "59": "Ranong",
-    "60": "Surat Thani",
-    "61": "Phangnga",
-    "62": "Phuket",
-    "63": "Krabi",
-    "64": "Nakhon Si Thammarat",
-    "65": "Trang",
-    "66": "Phatthalung",
-    "67": "Satun",
-    "68": "Songkhla",
-    "69": "Pattani",
-    "70": "Yala",
-    "71": "Ubon Ratchathani",
-    "72": "Yasothon",
-    "73": "Nakhon Phanom",
-    "74": "Prachin Buri",
-    "75": "Ubon Ratchathani",
-    "76": "Udon Thani",
-    "77": "Amnat Charoen",
-    "78": "Mukdahan",
-    "79": "Nong Bua Lamphu",
-    "80": "Sa Kaeo",
-    "81": "Bueng Kan"
-  },
-  "TJ": {
-    "1": "Kuhistoni Badakhshon",
-    "2": "Khatlon",
-    "3": "Sughd",
-    "4": "Dushanbe",
-    "5": "Nohiyahoi Tobei Jumhuri"
-  },
-  "TL": {
-    "6": "Dili"
-  },
-  "TM": {
-    "1": "Ahal",
-    "2": "Balkan",
-    "3": "Dashoguz",
-    "4": "Lebap",
-    "5": "Mary"
-  },
-  "TN": {
-    "2": "Kasserine",
-    "3": "Kairouan",
-    "6": "Jendouba",
-    "10": "Qafsah",
-    "14": "El Kef",
-    "15": "Al Mahdia",
-    "16": "Al Munastir",
-    "17": "Bajah",
-    "18": "Bizerte",
-    "19": "Nabeul",
-    "22": "Siliana",
-    "23": "Sousse",
-    "27": "Ben Arous",
-    "28": "Madanin",
-    "29": "Gabes",
-    "31": "Kebili",
-    "32": "Sfax",
-    "33": "Sidi Bou Zid",
-    "34": "Tataouine",
-    "35": "Tozeur",
-    "36": "Tunis",
-    "37": "Zaghouan",
-    "38": "Aiana",
-    "39": "Manouba"
-  },
-  "TO": {
-    "1": "Ha",
-    "2": "Tongatapu",
-    "3": "Vava"
-  },
-  "TR": {
-    "2": "Adiyaman",
-    "3": "Afyonkarahisar",
-    "4": "Agri",
-    "5": "Amasya",
-    "7": "Antalya",
-    "8": "Artvin",
-    "9": "Aydin",
-    "10": "Balikesir",
-    "11": "Bilecik",
-    "12": "Bingol",
-    "13": "Bitlis",
-    "14": "Bolu",
-    "15": "Burdur",
-    "16": "Bursa",
-    "17": "Canakkale",
-    "19": "Corum",
-    "20": "Denizli",
-    "21": "Diyarbakir",
-    "22": "Edirne",
-    "23": "Elazig",
-    "24": "Erzincan",
-    "25": "Erzurum",
-    "26": "Eskisehir",
-    "28": "Giresun",
-    "31": "Hatay",
-    "32": "Mersin",
-    "33": "Isparta",
-    "34": "Istanbul",
-    "35": "Izmir",
-    "37": "Kastamonu",
-    "38": "Kayseri",
-    "39": "Kirklareli",
-    "40": "Kirsehir",
-    "41": "Kocaeli",
-    "43": "Kutahya",
-    "44": "Malatya",
-    "45": "Manisa",
-    "46": "Kahramanmaras",
-    "48": "Mugla",
-    "49": "Mus",
-    "50": "Nevsehir",
-    "52": "Ordu",
-    "53": "Rize",
-    "54": "Sakarya",
-    "55": "Samsun",
-    "57": "Sinop",
-    "58": "Sivas",
-    "59": "Tekirdag",
-    "60": "Tokat",
-    "61": "Trabzon",
-    "62": "Tunceli",
-    "63": "Sanliurfa",
-    "64": "Usak",
-    "65": "Van",
-    "66": "Yozgat",
-    "68": "Ankara",
-    "69": "Gumushane",
-    "70": "Hakkari",
-    "71": "Konya",
-    "72": "Mardin",
-    "73": "Nigde",
-    "74": "Siirt",
-    "75": "Aksaray",
-    "76": "Batman",
-    "77": "Bayburt",
-    "78": "Karaman",
-    "79": "Kirikkale",
-    "80": "Sirnak",
-    "81": "Adana",
-    "82": "Cankiri",
-    "83": "Gaziantep",
-    "84": "Kars",
-    "85": "Zonguldak",
-    "86": "Ardahan",
-    "87": "Bartin",
-    "88": "Igdir",
-    "89": "Karabuk",
-    "90": "Kilis",
-    "91": "Osmaniye",
-    "92": "Yalova",
-    "93": "Duzce"
-  },
-  "TT": {
-    "1": "Arima",
-    "2": "Caroni",
-    "3": "Mayaro",
-    "4": "Nariva",
-    "5": "Port-of-Spain",
-    "6": "Saint Andrew",
-    "7": "Saint David",
-    "8": "Saint George",
-    "9": "Saint Patrick",
-    "10": "San Fernando",
-    "11": "Tobago",
-    "12": "Victoria"
-  },
-  "TW": {
-    "1": "Fu-chien",
-    "2": "Kao-hsiung",
-    "3": "T'ai-pei",
-    "4": "T'ai-wan"
-  },
-  "TZ": {
-    "2": "Pwani",
-    "3": "Dodoma",
-    "4": "Iringa",
-    "5": "Kigoma",
-    "6": "Kilimanjaro",
-    "7": "Lindi",
-    "8": "Mara",
-    "9": "Mbeya",
-    "10": "Morogoro",
-    "11": "Mtwara",
-    "12": "Mwanza",
-    "13": "Pemba North",
-    "14": "Ruvuma",
-    "15": "Shinyanga",
-    "16": "Singida",
-    "17": "Tabora",
-    "18": "Tanga",
-    "19": "Kagera",
-    "20": "Pemba South",
-    "21": "Zanzibar Central",
-    "22": "Zanzibar North",
-    "23": "Dar es Salaam",
-    "24": "Rukwa",
-    "25": "Zanzibar Urban",
-    "26": "Arusha",
-    "27": "Manyara"
-  },
-  "UA": {
-    "1": "Cherkas'ka Oblast'",
-    "2": "Chernihivs'ka Oblast'",
-    "3": "Chernivets'ka Oblast'",
-    "4": "Dnipropetrovs'ka Oblast'",
-    "5": "Donets'ka Oblast'",
-    "6": "Ivano-Frankivs'ka Oblast'",
-    "7": "Kharkivs'ka Oblast'",
-    "8": "Khersons'ka Oblast'",
-    "9": "Khmel'nyts'ka Oblast'",
-    "10": "Kirovohrads'ka Oblast'",
-    "11": "Krym",
-    "12": "Kyyiv",
-    "13": "Kyyivs'ka Oblast'",
-    "14": "Luhans'ka Oblast'",
-    "15": "L'vivs'ka Oblast'",
-    "16": "Mykolayivs'ka Oblast'",
-    "17": "Odes'ka Oblast'",
-    "18": "Poltavs'ka Oblast'",
-    "19": "Rivnens'ka Oblast'",
-    "20": "Sevastopol'",
-    "21": "Sums'ka Oblast'",
-    "22": "Ternopil's'ka Oblast'",
-    "23": "Vinnyts'ka Oblast'",
-    "24": "Volyns'ka Oblast'",
-    "25": "Zakarpats'ka Oblast'",
-    "26": "Zaporiz'ka Oblast'",
-    "27": "Zhytomyrs'ka Oblast'"
-  },
-  "UG": {
-    "26": "Apac",
-    "28": "Bundibugyo",
-    "29": "Bushenyi",
-    "30": "Gulu",
-    "31": "Hoima",
-    "33": "Jinja",
-    "36": "Kalangala",
-    "37": "Kampala",
-    "38": "Kamuli",
-    "39": "Kapchorwa",
-    "40": "Kasese",
-    "41": "Kibale",
-    "42": "Kiboga",
-    "43": "Kisoro",
-    "45": "Kotido",
-    "46": "Kumi",
-    "47": "Lira",
-    "50": "Masindi",
-    "52": "Mbarara",
-    "56": "Mubende",
-    "58": "Nebbi",
-    "59": "Ntungamo",
-    "60": "Pallisa",
-    "61": "Rakai",
-    "65": "Adjumani",
-    "66": "Bugiri",
-    "67": "Busia",
-    "69": "Katakwi",
-    "70": "Luwero",
-    "71": "Masaka",
-    "72": "Moyo",
-    "73": "Nakasongola",
-    "74": "Sembabule",
-    "76": "Tororo",
-    "77": "Arua",
-    "78": "Iganga",
-    "79": "Kabarole",
-    "80": "Kaberamaido",
-    "81": "Kamwenge",
-    "82": "Kanungu",
-    "83": "Kayunga",
-    "84": "Kitgum",
-    "85": "Kyenjojo",
-    "86": "Mayuge",
-    "87": "Mbale",
-    "88": "Moroto",
-    "89": "Mpigi",
-    "90": "Mukono",
-    "91": "Nakapiripirit",
-    "92": "Pader",
-    "93": "Rukungiri",
-    "94": "Sironko",
-    "95": "Soroti",
-    "96": "Wakiso",
-    "97": "Yumbe"
-  },
-  "US": {
-    "848": "Armed Forces Americas",
-    "852": "Armed Forces Europe",
-    "858": "Alaska",
-    "859": "Alabama",
-    "863": "Armed Forces Pacific",
-    "865": "Arkansas",
-    "866": "American Samoa",
-    "873": "Arizona",
-    "934": "California",
-    "948": "Colorado",
-    "953": "Connecticut",
-    "979": "District of Columbia",
-    "981": "Delaware",
-    "1074": "Florida",
-    "1075": "Federated States of Micronesia",
-    "1106": "Georgia",
-    "1126": "Guam",
-    "1157": "Hawaii",
-    "1192": "Iowa",
-    "1195": "Idaho",
-    "1203": "Illinois",
-    "1205": "Indiana",
-    "1296": "Kansas",
-    "1302": "Kentucky",
-    "1321": "Louisiana",
-    "1364": "Massachusetts",
-    "1367": "Maryland",
-    "1368": "Maine",
-    "1371": "Marshall Islands",
-    "1372": "Michigan",
-    "1377": "Minnesota",
-    "1378": "Missouri",
-    "1379": "Northern Mariana Islands",
-    "1382": "Mississippi",
-    "1383": "Montana",
-    "1409": "North Carolina",
-    "1410": "North Dakota",
-    "1411": "Nebraska",
-    "1414": "New Hampshire",
-    "1416": "New Jersey",
-    "1419": "New Mexico",
-    "1428": "Nevada",
-    "1431": "New York",
-    "1457": "Ohio",
-    "1460": "Oklahoma",
-    "1467": "Oregon",
-    "1493": "Pennsylvania",
-    "1515": "Palau",
-    "1587": "Rhode Island",
-    "1624": "South Carolina",
-    "1625": "South Dakota",
-    "1678": "Tennessee",
-    "1688": "Texas",
-    "1727": "Utah",
-    "1751": "Virginia",
-    "1759": "Virgin Islands",
-    "1770": "Vermont",
-    "1794": "Washington",
-    "1802": "Wisconsin",
-    "1815": "West Virginia",
-    "1818": "Wyoming"
-  },
-  "UY": {
-    "1": "Artigas",
-    "2": "Canelones",
-    "3": "Cerro Largo",
-    "4": "Colonia",
-    "5": "Durazno",
-    "6": "Flores",
-    "7": "Florida",
-    "8": "Lavalleja",
-    "9": "Maldonado",
-    "10": "Montevideo",
-    "11": "Paysandu",
-    "12": "Rio Negro",
-    "13": "Rivera",
-    "14": "Rocha",
-    "15": "Salto",
-    "16": "San Jose",
-    "17": "Soriano",
-    "18": "Tacuarembo",
-    "19": "Treinta y Tres"
-  },
-  "UZ": {
-    "1": "Andijon",
-    "2": "Bukhoro",
-    "3": "Farghona",
-    "4": "Jizzakh",
-    "5": "Khorazm",
-    "6": "Namangan",
-    "7": "Nawoiy",
-    "8": "Qashqadaryo",
-    "9": "Qoraqalpoghiston",
-    "10": "Samarqand",
-    "11": "Sirdaryo",
-    "12": "Surkhondaryo",
-    "13": "Toshkent",
-    "14": "Toshkent",
-    "15": "Jizzax"
-  },
-  "VC": {
-    "1": "Charlotte",
-    "2": "Saint Andrew",
-    "3": "Saint David",
-    "4": "Saint George",
-    "5": "Saint Patrick",
-    "6": "Grenadines"
-  },
-  "VE": {
-    "1": "Amazonas",
-    "2": "Anzoategui",
-    "3": "Apure",
-    "4": "Aragua",
-    "5": "Barinas",
-    "6": "Bolivar",
-    "7": "Carabobo",
-    "8": "Cojedes",
-    "9": "Delta Amacuro",
-    "11": "Falcon",
-    "12": "Guarico",
-    "13": "Lara",
-    "14": "Merida",
-    "15": "Miranda",
-    "16": "Monagas",
-    "17": "Nueva Esparta",
-    "18": "Portuguesa",
-    "19": "Sucre",
-    "20": "Tachira",
-    "21": "Trujillo",
-    "22": "Yaracuy",
-    "23": "Zulia",
-    "24": "Dependencias Federales",
-    "25": "Distrito Federal",
-    "26": "Vargas"
-  },
-  "VN": {
-    "1": "An Giang",
-    "3": "Ben Tre",
-    "5": "Cao Bang",
-    "9": "Dong Thap",
-    "13": "Hai Phong",
-    "20": "Ho Chi Minh",
-    "21": "Kien Giang",
-    "23": "Lam Dong",
-    "24": "Long An",
-    "30": "Quang Ninh",
-    "32": "Son La",
-    "33": "Tay Ninh",
-    "34": "Thanh Hoa",
-    "35": "Thai Binh",
-    "37": "Tien Giang",
-    "39": "Lang Son",
-    "43": "Dong Nai",
-    "44": "Ha Noi",
-    "45": "Ba Ria-Vung Tau",
-    "46": "Binh Dinh",
-    "47": "Binh Thuan",
-    "49": "Gia Lai",
-    "50": "Ha Giang",
-    "52": "Ha Tinh",
-    "53": "Hoa Binh",
-    "54": "Khanh Hoa",
-    "55": "Kon Tum",
-    "58": "Nghe An",
-    "59": "Ninh Binh",
-    "60": "Ninh Thuan",
-    "61": "Phu Yen",
-    "62": "Quang Binh",
-    "63": "Quang Ngai",
-    "64": "Quang Tri",
-    "65": "Soc Trang",
-    "66": "Thua Thien-Hue",
-    "67": "Tra Vinh",
-    "68": "Tuyen Quang",
-    "69": "Vinh Long",
-    "70": "Yen Bai",
-    "71": "Bac Giang",
-    "72": "Bac Kan",
-    "73": "Bac Lieu",
-    "74": "Bac Ninh",
-    "75": "Binh Duong",
-    "76": "Binh Phuoc",
-    "77": "Ca Mau",
-    "78": "Da Nang",
-    "79": "Hai Duong",
-    "80": "Ha Nam",
-    "81": "Hung Yen",
-    "82": "Nam Dinh",
-    "83": "Phu Tho",
-    "84": "Quang Nam",
-    "85": "Thai Nguyen",
-    "86": "Vinh Phuc",
-    "87": "Can Tho",
-    "88": "Dac Lak",
-    "89": "Lai Chau",
-    "90": "Lao Cai",
-    "91": "Dak Nong",
-    "92": "Dien Bien",
-    "93": "Hau Giang"
-  },
-  "VU": {
-    "5": "Ambrym",
-    "6": "Aoba",
-    "7": "Torba",
-    "8": "Efate",
-    "9": "Epi",
-    "10": "Malakula",
-    "11": "Paama",
-    "12": "Pentecote",
-    "13": "Sanma",
-    "14": "Shepherd",
-    "15": "Tafea",
-    "16": "Malampa",
-    "17": "Penama",
-    "18": "Shefa"
-  },
-  "WS": {
-    "2": "Aiga-i-le-Tai",
-    "3": "Atua",
-    "4": "Fa",
-    "5": "Gaga",
-    "6": "Va",
-    "7": "Gagaifomauga",
-    "8": "Palauli",
-    "9": "Satupa",
-    "10": "Tuamasaga",
-    "11": "Vaisigano"
-  },
-  "YE": {
-    "1": "Abyan",
-    "2": "Adan",
-    "3": "Al Mahrah",
-    "4": "Hadramawt",
-    "5": "Shabwah",
-    "6": "Lahij",
-    "7": "Al Bayda'",
-    "8": "Al Hudaydah",
-    "9": "Al Jawf",
-    "10": "Al Mahwit",
-    "11": "Dhamar",
-    "12": "Hajjah",
-    "13": "Ibb",
-    "14": "Ma'rib",
-    "15": "Sa'dah",
-    "16": "San'a'",
-    "17": "Taizz",
-    "18": "Ad Dali",
-    "19": "Amran",
-    "20": "Al Bayda'",
-    "21": "Al Jawf",
-    "22": "Hajjah",
-    "23": "Ibb",
-    "24": "Lahij",
-    "25": "Taizz"
-  },
-  "ZA": {
-    "1": "North-Western Province",
-    "2": "KwaZulu-Natal",
-    "3": "Free State",
-    "5": "Eastern Cape",
-    "6": "Gauteng",
-    "7": "Mpumalanga",
-    "8": "Northern Cape",
-    "9": "Limpopo",
-    "10": "North-West",
-    "11": "Western Cape"
-  },
-  "ZM": {
-    "1": "Western",
-    "2": "Central",
-    "3": "Eastern",
-    "4": "Luapula",
-    "5": "Northern",
-    "6": "North-Western",
-    "7": "Southern",
-    "8": "Copperbelt",
-    "9": "Lusaka"
-  },
-  "ZW": {
-    "1": "Manicaland",
-    "2": "Midlands",
-    "3": "Mashonaland Central",
-    "4": "Mashonaland East",
-    "5": "Mashonaland West",
-    "6": "Matabeleland North",
-    "7": "Matabeleland South",
-    "8": "Masvingo",
-    "9": "Bulawayo",
-    "10": "Harare"
-  }
-};
