--- legacy/lib/region.js
@@ -1,6 +1,0 @@
-
-module.exports = function() {
-  this.countryCode = null;
-  this.countryName = null;
-  this.region = null;
-};
