--- legacy/lib/time_zone.js
@@ -1,9 +1,0 @@
-var _data = require('./data/time_zone_data');
-
-module.exports = function(country, region) {
-  if (!country)
-    return null;
-
-  var result = _data[country];
-  return result ? result[region] || result[''] : null;
-};
