--- legacy/test/dbs/test.dat
@@ -1,1 +1,0 @@
-abcdef
