--- legacy/test/index.js
@@ -1,4 +1,0 @@
-var assert = require('assert');
-
-// Enable highlighted diffs in mocha
-assert.AssertionError.prototype.showDiff = true;
