--- legacy/tools/gen_continents.rb
@@ -1,25 +1,0 @@
-require "open-uri"
-require "json"
-
-
-raw_data = open("http://dev.maxmind.com/static/csv/codes/country_continent.csv").read
-res = {}
-
-# US,NA
-data = raw_data.split("\n").map do |line|
-  row = line.gsub(/\"/, "").split(",")
-  next if row[0] == "iso 3166 country"
-  res[row[0]] = {} unless res.has_key?(row[0])
-  res[row[0]]= row[1];
-end
-
-continents = File.open(File.join(File.dirname(__FILE__), "/../lib/data/continents.js"), "w")
-continents.write %Q{
-/* jshint indent: 2 */
-// Do not modify this file, it was generated by
-// ./tools/gen_continents.rb
-module.exports = #{JSON.pretty_generate(res)};
-}
-continents.close
-
-puts "Done"
