--- legacy/tools/gen_time_zone.rb
@@ -1,25 +1,0 @@
-require "open-uri"
-require "json"
-
-
-raw_data = open("http://www.maxmind.com/timezone.txt").read
-res = {}
-
-# US,AL,"America/Chicago"
-data = raw_data.split("\n").map do |line|
-  row = line.gsub(/\"/, "").split(",")
-  next if row[0] == "country"
-  res[row[0]] = {} unless res.has_key?(row[0])
-  res[row[0]][row[1]] = row[2]
-end
-
-time_zone_js = File.open(File.join(File.dirname(__FILE__), "/../lib/data/time_zone_data.js"), "w")
-time_zone_js.write %Q{
-/* jshint indent: 2 */
-// Do not modify this file, it was generated by
-// ./tools/gen_time_zone.rb
-module.exports = #{JSON.pretty_generate(res)};
-}
-time_zone_js.close
-
-puts "Done"
