+++ lib/Reader.js
@@ -1,9 +1,10 @@
-var DATA_SECTION_SEPARATOR_SIZE = 16,
-  METADATA_START_MARKER = new Buffer('ABCDEF4D61784D696E642E636F6D', 'hex'),
-  fs = require('fs'),
-  Metadata = require('./Metadata.js'),
-  Decoder = require('./Decoder.js'),
-  IPParser = require('./IPParser.js');
+var fs = require('fs');
+var Metadata = require('./metadata');
+var Decoder = require('./decoder');
+var IPParser = require('./ip');
+
+const DATA_SECTION_SEPARATOR_SIZE = 16;
+const METADATA_START_MARKER = new Buffer('ABCDEF4D61784D696E642E636F6D', 'hex');
 
 function concat2(a, b) {
   return (a << 8) | b;
